+++ paper/paper.bib
@@ -4,7 +4,8 @@
   journal = {Proceedings of the Royal Society of London},
   volume = {58},
   pages = {240--242},
-  year = {1895}
+  year = {1895},
+  doi = {10.1098/rspl.1895.0041}
 }
 
 @article{szekely:2007,
@@ -15,7 +16,8 @@
   number = {6},
   pages = {2769--2794},
   year = {2007},
-  publisher = {Institute of Mathematical Statistics}
+  publisher = {Institute of Mathematical Statistics},
+  doi = {10.1214/009053607000000505}
 }
 
 @article{reshef:2011,
@@ -26,7 +28,8 @@
   number = {6062},
   pages = {1518--1524},
   year = {2011},
-  publisher = {American Association for the Advancement of Science}
+  publisher = {American Association for the Advancement of Science},
+  doi = {10.1126/science.1205438}
 }
 
 @inproceedings{flach:2012,
@@ -36,7 +39,8 @@
   year = {2012},
   editor = {Pereira, F. and Burges, C.J.C. and Bottou, L. and Weinberger, K.Q.},
   publisher = {Curran Associates, Inc.},
-  pages = {838--846}
+  pages = {838--846},
+  doi = {10.5555/2999134.2999236}
 }
 
 @book{cohen:1983,
@@ -44,21 +48,24 @@
   title = {Applied Multiple Regression/Correlation Analysis for the Behavioral Sciences},
   year = {1983},
   edition = {2nd},
-  publisher = {Lawrence Erlbaum Associates}
+  publisher = {Lawrence Erlbaum Associates},
+  doi = {10.4324/9780203774441}
 }
 
 @article{cramer:1946,
   author = {Cramér, Harald},
   title = {Mathematical Methods of Statistics},
   journal = {Princeton University Press},
-  year = {1946}
+  year = {1946},
+  doi = {10.1515/9781400883868}
 }
 
 @book{theil:1970,
   author = {Theil, Henri},
   title = {Statistical Decomposition Analysis: With Applications in the Social and Administrative Sciences},
   year = {1970},
-  publisher = {North-Holland Publishing Company}
+  publisher = {North-Holland Publishing Company},
+  doi = {10.1016/C2013-0-07379-0}
 }
 
 @article{tibshirani:2010,
@@ -69,6 +76,6 @@
   number = {2},
   pages = {411--423},
   year = {2001},
-  publisher = {Wiley Online Library}
+  publisher = {Wiley Online Library},
+  doi = {10.1111/1467-9868.00293}
 }
-
