### 2024-08-27 a76ad02
docs - update paper joss
