### 2024-08-30 #7143 @crvernon
:wave:  @PHS-Meantrix - thanks for your submission to JOSS.  I am going to run this one by our larger editorial board due to it's smaller code base to ensure it meets our [substantial scholarly effort](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort) requirement.  This should take about two weeks at the most.  I'll let you know what I hear back ASAP.
