### 2024-09-01 #7143 @PHS-Meantrix
Dear JOSS Editorial Team,

Since it was our first submission, I initially thought there would be
another step to review the PDF before finalizing the submission, but there
wasn't one. After learning how to review the PDF and checking it, we’ve
noticed that some tables need formatting, along with minor text updates and
corrections to the DOIs in the references. We recently committed updates to
the `paper.md`. How can we proceed with updating the submission?

Best regards,
Paulo

Em sex., 30 de ago. de 2024 às 09:34, The Open Journals editorial robot <
***@***.***> escreveu:

> Submission flagged for editorial review.
>
