### 2024-09-02 #7143 @danielskatz
@PHS-Meantrix - you can freely edit the submission, at least before the review starts.

Also, bot commands you can use are `@editorialbot check references` (to check the references/DOIs), and `@editorialbot generate pdf` (to make a new PDF). editorialbot commands need to be the first entry in a new comment.
