### 2024-09-06 1b8e6e9
docs - change bib refs
