+++ paper/paper.bib
@@ -9,7 +9,7 @@
 }
 
 @article{szekely:2007,
-  author = {Székely, Gábor J. and Rizzo, Maria L. and Bakirov, Nail K.},
+  author = {Székely,  Gábor J. and Rizzo,  Maria L. and Bakirov,  Nail K.},
   title = {Measuring and testing independence by correlation of distances},
   journal = {The Annals of Statistics},
   volume = {35},
@@ -32,16 +32,16 @@
   doi = {10.1126/science.1205438}
 }
 
-@inproceedings{flach:2012,
-  author = {Flach, Peter A. and Kull, Meelis and G\"{u}venir, H. Altay},
-  title = {Precision-Recall-Gain Curves: PR Analysis Done Right},
-  booktitle = {Advances in Neural Information Processing Systems 25},
-  year = {2012},
-  editor = {Pereira, F. and Burges, C.J.C. and Bottou, L. and Weinberger, K.Q.},
-  publisher = {Curran Associates, Inc.},
-  pages = {838--846},
-  doi = {10.5555/2999134.2999236}
+@misc{pps:2020,
+  doi = {10.5281/ZENODO.4091345},
+  url = {https://zenodo.org/record/4091345},
+  author = {Wetschoreck,  Florian and Krabel,  Tobias and Krishnamurthy,  Surya},
+  title = {8080labs/ppscore: zenodo release},
+  publisher = {Zenodo},
+  year = {2020},
+  copyright = {Open Access}
 }
+
 
 @book{cohen:1983,
   author = {Cohen, Jacob},
@@ -68,14 +68,15 @@
   doi = {10.1016/C2013-0-07379-0}
 }
 
-@article{tibshirani:2010,
-  author = {Tibshirani, Ryan J. and Walther, Guenther and Hastie, Trevor},
-  title = {Estimating the number of clusters in a data set via the gap statistic},
-  journal = {Journal of the Royal Statistical Society: Series B (Statistical Methodology)},
-  volume = {63},
-  number = {2},
-  pages = {411--423},
-  year = {2001},
-  publisher = {Wiley Online Library},
-  doi = {10.1111/1467-9868.00293}
+@article{bhattacharya:2010,
+title = {Average correlation clustering algorithm (ACCA) for grouping of co-regulated genes with similar pattern of variation in their expression values},
+journal = {Journal of Biomedical Informatics},
+volume = {43},
+number = {4},
+pages = {560-568},
+year = {2010},
+issn = {1532-0464},
+doi = {https://doi.org/10.1016/j.jbi.2010.02.001},
+url = {https://www.sciencedirect.com/science/article/pii/S1532046410000158},
+author = {Anindya Bhattacharya and Rajat K. De}
 }
