### 2024-09-06 bd32fa3
docs - refined article text JOSS
