+++ paper/paper.md
@@ -24,40 +24,35 @@ bibliography: paper.bib
 
 # Summary
 
-Correlation-like analysis provides an important statistical measure that can describe the size and direction of an association between variables. However, there are few R packages that can efficiently perform this type of analysis on large datasets with mixed data types. The **corrp** package provides a full suite of solutions for computing various kinds of correlation-like measures such as Pearson correlation [@pearson:1895], Distance Correlation [@szekely:2007], Maximal Information Coefficient (MIC) [@reshef:2011], Predictive Power Score (PPS) [@flach:2012], Cramér's V [@cramer:1946], and the Uncertainty Coefficient [@theil:1970]. These methods support the analysis of data frames with mixed classes (integer, numeric, factor, and character). In addition, it offers a C++ implementation of the Average Correlation Clustering Algorithm (ACCA) [@tibshirani:2010], which is original used in genetic studies. In this package, the ACCA algorithm has been extended to work directly with correlation matrices derived from different association methods, depending on the data types and user preferences. Furthermore, the package is designed for parallel processing in R, making it highly efficient for large datasets.
+Correlation-like analysis provides an important statistical measure that can describe the size and direction of an association between variables. However, there are few R packages that can efficiently perform this type of analysis on large datasets with mixed data types. The `corrp` package provides a full suite of solutions for computing various kinds of correlation-like measures such as Pearson correlation [@pearson:1895], Distance Correlation [@szekely:2007], Maximal Information Coefficient (MIC) [@reshef:2011], Predictive Power Score (PPS) [@pps:2020], Cramér's V [@cramer:1946], and the Uncertainty Coefficient [@theil:1970]. These methods support the analysis of data frames with mixed classes (integer, numeric, factor, and character). In addition, it offers a C++ implementation of the Average Correlation Clustering Algorithm (ACCA) [@bhattacharya:2010], which is original used in genetic studies. In this package, the ACCA algorithm has been extended to work directly with correlation matrices derived from different association methods, depending on the data types and user preferences. Furthermore, the package is designed for parallel processing in R, making it highly efficient for large datasets.
 
 
 
 # Statement of need
 
-The corrp package is an R package that provides a flexible and efficient way of doing correlation-like analysis on mixed-type data frames. Most traditional correlation methods in R are applicable only to specific data types or small datasets.**corrp** extends this capability by handling mixed classes, integrating various association methods, and offering clustering directly from the resulting correlation matrix. The package is particularly useful for researchers and data scientists working with complex datasets who require robust and scalable tools for both association analysis and clustering.
+The corrp package is an R package that provides a flexible and efficient way of doing correlation-like analysis on mixed-type data frames. Most traditional correlation methods in R are applicable only to specific data types or small datasets `corrp` extends this capability by handling mixed classes, integrating various association methods, and offering clustering directly from the resulting correlation matrix. The package is particularly useful for researchers and data scientists working with complex datasets who require robust and scalable tools for both association analysis and clustering.
 
 
 # Implementation
 
-R and C++ were combined in the creation of corrp package to merge flexibility of R with the speed of C++ which is used for key operations. The core functionalities include:
[…]
