+++ paper/paper.bib
@@ -76,7 +76,7 @@ number = {4},
 pages = {560-568},
 year = {2010},
 issn = {1532-0464},
-doi = {https://doi.org/10.1016/j.jbi.2010.02.001},
+doi = {10.1016/j.jbi.2010.02.001},
 url = {https://www.sciencedirect.com/science/article/pii/S1532046410000158},
 author = {Anindya Bhattacharya and Rajat K. De}
 }
