+++ paper/paper.bib
@@ -60,12 +60,11 @@
   doi = {10.1515/9781400883868}
 }
 
-@book{theil:1970,
+@book{theil:1972,
   author = {Theil, Henri},
   title = {Statistical Decomposition Analysis: With Applications in the Social and Administrative Sciences},
-  year = {1970},
-  publisher = {North-Holland Publishing Company},
-  doi = {10.1016/C2013-0-07379-0}
+  year = {1972},
+  publisher = {North-Holland Publishing Company}
 }
 
 @article{bhattacharya:2010,
