### 2024-09-07 #7143 @PHS-Meantrix
@danielskatz Thanks, another thing we want to know how to change the citation format for our author names. Instead of "Siciliani, & dos Santos," can we change to Igor D.S. Siciliani, & P.H. Santos?

Also, there is no doi for Henri Theil's book. How to proceed with this?
