### 2024-09-07 #7143 @danielskatz
The citation format can't be changed, unless there is something incorrect in the parsing of your names, which doesn't appear to be the case here.

And for items with no DOI, that's fine - you don't need to do anything.
