### 2024-09-10 #7143 @jromanowska
@PHS-Meantrix , before I start, I have one question. In the acknowledgments you write that you took inspiration for the package from the code written by Srikanth Komala Sheshachala. Could you specify if there are any parts of the code that are copied from there? I can't see any licence in their repo (https://github.com/talegari/sidekicks/), so if you have used their code, we should probably invite them as co-author.
