### 2024-10-07 #7319 @editorialbot
[REVIEW]: corrp: An R package for multiple correlation-like analysis and clustering in mixed data
**Submitting author:** @PHS-Meantrix (<a href="http://orcid.org/0009-0004-5273-4142">Paulo Henrique dos Santos</a>)
**Repository:** https://github.com/meantrix/corrp
**Branch with paper.md** (empty if default branch): JOSS
**Version:** 0.4.0
