### 2024-10-08 26f0deb
add energy to suggests for docs
