### 2024-10-08 meantrix/corrp#8 @malcolmbarrett
Authorship clarification
In openjournals/joss-reviews#7319, @PHS-Meantrix is listed as the submitting author and an author on the paper. In the `DESCRIPTION` file, @igor-siciliani together with `Meantrix <dev@meantrix.com>`. Should @PHS-Meantrix be an author or contributed? I also believe that Meantrix should be the copyright holder, not the maintainer. See this example from Posit of how to list authors and companies https://github.com/r-lib/usethis/blob/main/DESCRIPTION
