### 2024-10-08 meantrix/corrp#9 @malcolmbarrett
Submitting to CRAN?
I was surprised this wasn't first submitted to CRAN which would have likely addressed several issues already presented.

Are you planning to submit to CRAN? If not, I think you should clarify that in the README. If so, I would suggest you submit sooner than later to better time it with the paper.

openjournals/joss-reviews#7319
