### 2024-10-08 meantrix/corrp#10 @malcolmbarrett
Add package website and longer form documentation
You should have a package website for documentation and the like, made with something like pkgdown or altdoc.

You can do this very quickly on GitHub with `usethis::use_pkgdown_github_pages()`

I would also like to see a vignette (`usethis::use_vignette()`) or article (`usethis::use_article()`) with the functionality of the package discussed, similar to the JOSS paper but made for new users or covering advanced topics

openjournals/joss-reviews#7319
