### 2024-10-08 meantrix/corrp#13 @malcolmbarrett
Document performance and benchmarks
The README and the JOSS paper make claims about the performance of this package. I think you need to substantiate those claims with performance benchmarks.

I also think you need to give users some guidance as to what to expect and what tools take longer than others. I was very surprised how long it took to run `results = corrp::corrp(iris, cor.nn = 'mic',cor.nc = 'pps',cor.cc = 'uncoef', n.cores = 2 , verbose = FALSE)` given the claims of the package and the use of 2 cores for a dataset as small as iris. If that's expected based on the type of calculations being done, that needs to be clearer.

A more realistic dataset might help, as well

openjournals/joss-reviews#7319
