+++ DESCRIPTION
@@ -25,6 +25,7 @@ Imports:
     ppsr (>= 0.0.2),
     DescTools(>= 0.99.40)
 Suggests: 
+    corrplot,
     energy,
     testthat
 License: GPL (>= 3)
