### 2024-10-09 meantrix/corrp#8 @jromanowska
Very good point! However, in JOSS, the authors on the paper should be those who contributed to the code in the repo. Thus, the author list is fine. Let's see what the authors say about the items in `DESCRIPTION` .
