### 2024-10-09 #7319 @jromanowska
Thank you for the comments and issues, @malcolmbarrett. Great work so far!
It's important that the packages that are published in JOSS follow the guidelines and good practices of their language. Therefore, the authors should work on the points you've raised and we'll see where does it take us.
The authors should check their package, e.g., with `R CMD check`. However, we do not require CRAN submission as this is very strict.

Please notify the authors whenever you have a task/comment for them. @PHS-Meantrix
