### 2024-10-09 #7319 @PHS-Meantrix
In the coming days, as suggested by @malcolmbarrett, we will be adjust this package to the CRAN format. We will also review and work towards resolving the issues that have been opened in the package. Thank you for the comments.
