+++ tests/testthat/test-corrp.R
@@ -1,170 +1,162 @@
 test_that("Tests on corrp and cor_fun functions", {
-
-  #data to tests
-  df = iris
-
-
-  #class test
-  corr = corrp(df,comp = 'g', alternative = 't')
-  expect_s3_class(corr,'clist')
-
-  #parallel test
-  corr2 = corrp(df,comp = 'g', alternative = 't',parallel = F)
-  expect_s3_class(corr2,'clist')
-  expect_equal(corr,corr2)
-
-  #P-Value test
-  data = corr$data
-  data.p1 = subset(data,stat.value < 0.05 )
-  expect_true( all(data.p1$isig) )
+  # data to tests
+  df <- iris
 
 
-  #Inferences test 1: corrp
-  data.pearson = subset(data,infer == "Pearson Correlation")
-  data.cramer = subset(data,infer == "Cramer's V")
-  data.linear = subset(data,infer == "Linear Model")
+  # class test
+  corr <- corrp(df, comp = "g", alternative = "t")
+  expect_s3_class(corr, "clist")
 
-  #Pearson Correlation
-  for(i in 1:NROW(data.pearson)){
+  # parallel test
+  corr2 <- corrp(df, comp = "g", alternative = "t", parallel = F)
+  expect_s3_class(corr2, "clist")
+  expect_equal(corr, corr2)
 
-  i.test = stats::cor.test(df[[ as.character(data.pearson$varx[i]) ]], df[[ as.character(data.pearson$vary[i]) ]], alternative = 't')
+  # P-Value test
+  data <- corr$data
+  data.p1 <- subset(data, stat.value < 0.05)
+  expect_true(all(data.p1$isig))
 
-  expect_equal( i.test$p.value,data.pearson$stat.value[i] )
-  expect_equal( as.numeric(i.test$'estimate'), data.pearson$infer.value[i])
 
+  # Inferences test 1: corrp
+  data.pearson <- subset(data, infer == "Pearson Correlation")
+  data.cramer <- subset(data, infer == "Cramer's V")
+  data.linear <- subset(data, infer == "Linear Model")
 
+  # Pearson Correlation
+  for (i in 1:NROW(data.pearson)) {
+    i.test <- stats::cor.test(df[[as.character(data.pearson$varx[i])]], df[[as.character(data.pearson$vary[i])]], alternative = "t")
+
+    expect_equal(i.test$p.value, data.pearson$stat.value[i])
+    expect_equal(as.numeric(i.test$"estimate"), data.pearson$infer.value[i])
   }
 
-  #Cramer's V
-  for(i in 1:NROW(data.cramer)){
+  # Cramer's V
+  for (i in 1:NROW(data.cramer)) {
+    i.pv <- stats::chisq.test(df[[as.character(data.cramer$varx[i])]], df[[as.character(data.cramer$vary[i])]],
+      simulate.p.value = TRUE
+    )$p.value
+    i.test <- lsr::cramersV(iris[[as.character(data.cramer$varx[i])]], iris[[as.character(data.cramer$vary[i])]])
 
-    i.pv = stats::chisq.test(df[[ as.character(data.cramer$varx[i]) ]], df[[ as.character(data.cramer$vary[i]) ]],
-                             simulate.p.value=TRUE)$p.value
-    i.test = lsr::cramersV(iris[[ as.character(data.cramer$varx[i]) ]], iris[[ as.character(data.cramer$vary[i]) ]])
-
-    expect_equal( i.pv,data.cramer$stat.value[i] )
-    expect_equal( i.test, data.pearson$infer.value[i])
-
-
+    expect_equal(i.pv, data.cramer$stat.value[i])
[…]
