+++ tests/testthat/test-corr_rm.R
@@ -1,13 +1,11 @@
 test_that("Tests on corr_rm", {
+  # data to tests
+  df <- iris
 
-  #data to tests
-  df = iris
+  corr <- corrp(df, comp = "g", alternative = "t")
+  m <- corr_matrix(corr)
+  df2 <- corr_rm(df = df, c = corr)
+  df3 <- corr_rm(df = df, c = m)
 
-  corr = corrp(df,comp = 'g', alternative = 't')
-  m = corr_matrix(corr)
-  df2 = corr_rm(df =df,c = corr)
-  df3 = corr_rm(df =df,c = m)
-
-  expect_equal(df2,df3)
-
+  expect_equal(df2, df3)
 })
