+++ R/utils.R
@@ -1,34 +1,34 @@
 # auxiliar functions
 
 # linear regression Calculations
-.corlm = function(x, y, nx, ny, p.value, comp, verbose, lm.args = list(), ...) {
-  if (is.data.frame(x)) x = x[[1]]
-  if (is.data.frame(y)) y = y[[1]]
+.corlm <- function(x, y, nx, ny, p.value, comp, verbose, lm.args = list(), ...) {
+  if (is.data.frame(x)) x <- x[[1]]
+  if (is.data.frame(y)) y <- y[[1]]
 
-  infer = "Linear Model"
-  stat = "P-value"
+  infer <- "Linear Model"
+  stat <- "P-value"
 
-  args = c(list(y ~ as.factor(x)), lm.args)
+  args <- c(list(y ~ as.factor(x)), lm.args)
 
-  sum.res = summary(
+  sum.res <- summary(
     do.call(stats::lm, args)
   )
 
-  pv = stats::pf(sum.res$fstatistic[1], sum.res$fstatistic[2],
+  pv <- stats::pf(sum.res$fstatistic[1], sum.res$fstatistic[2],
     sum.res$fstatistic[3],
     lower.tail = F
   )
 
-  compare = .comparepv(x = pv, pv = p.value, comp = comp)
-  r = sqrt(sum.res[["r.squared"]])
-  msg = ""
+  compare <- .comparepv(x = pv, pv = p.value, comp = comp)
+  r <- sqrt(sum.res[["r.squared"]])
+  msg <- ""
 
 
   if (compare$comp) {
-    isig = TRUE
+    isig <- TRUE
 
     if (verbose) {
-      msg = paste(
+      msg <- paste(
         ny, "vs.", nx, ".",
         "Alternative hypothesis: true ", infer, " is not equal to 0.",
         "P-value: ", pv, "."
@@ -37,10 +37,10 @@
       message(msg)
     }
   } else {
-    isig = FALSE
+    isig <- FALSE
 
     if (verbose) {
-      msg = paste(
+      msg <- paste(
         ny, "vs.", nx, ".",
         "There is no correlation at the confidence level  p-value.",
         "P-value:", p.value, compare$str, "estimated p-value:", pv
@@ -58,26 +58,26 @@
 }
 
 # CramersV Calculations
-.cramersvp = function(x, y, nx, ny, p.value, comp, verbose, cramersV.args = list(), ...) {
-  if (is.data.frame(x)) x = x[[1]]
-  if (is.data.frame(y)) y = y[[1]]
+.cramersvp <- function(x, y, nx, ny, p.value, comp, verbose, cramersV.args = list(), ...) {
+  if (is.data.frame(x)) x <- x[[1]]
+  if (is.data.frame(y)) y <- y[[1]]
 
-  infer = "Cramer's V"
-  stat = "P-value"
+  infer <- "Cramer's V"
+  stat <- "P-value"
 
-  args = c(list(x), list(y), cramersV.args)
+  args <- c(list(x), list(y), cramersV.args)
 
-  pv = stats::chisq.test(x, y, simulate.p.value = TRUE)$p.value
-  r = do.call(lsr::cramersV, args)
-  compare = .comparepv(x = pv, pv = p.value, comp = comp)
-  msg = ""
+  pv <- stats::chisq.test(x, y, simulate.p.value = TRUE)$p.value
+  r <- do.call(lsr::cramersV, args)
+  compare <- .comparepv(x = pv, pv = p.value, comp = comp)
+  msg <- ""
 
 
   if (compare$comp) {
-    isig = TRUE
+    isig <- TRUE
 
     if (verbose) {
-      msg = paste(
+      msg <- paste(
[…]

[…]
