+++ R/acca.R
@@ -1,16 +1,17 @@
 #' @useDynLib corrp, .registration=TRUE
-#' @exportPattern "^[[:alpha:]]+"
 #' @importFrom Rcpp evalCpp
 #' @importFrom RcppArmadillo armadillo_version
 
 #' @title Average correlation clustering algorithm
 #'
 #' @description A C++ implementation of the ACCA method
-#' that works directly with the correlation matrix derived from the \code{\link{corr_matrix}} function.
+#' that works directly with the correlation 
+#' matrix derived from the \code{\link{corr_matrix}} function.
 #' In this sense, this implementation differs from the original,
 #' it works with mixed data and several correlation methods.
 #'
-#' @param m  \[\code{matrix(1)}]\cr correlation matrix from \code{\link{corr_matrix}} or a distance matrix.
+#' @param m  \[\code{matrix(1)}]\cr correlation matrix from 
+#' \code{\link{corr_matrix}} or a distance matrix.
 #' @param k \[\code{integer(1)}]\cr number of clusters considered.
 #' @param maxrep \[\code{integer(1)}]\cr maximum number of interactions without change in the clusters.
 #' @param maxiter \[\code{integer(1)}]\cr maximum number of interactions.
