+++ tests/testthat/test-corr_matrix.R
@@ -8,11 +8,11 @@ test_that("Tests on corr_matrix", {
   # class test
   expect_s3_class(m, "cmatrix")
 
-  si <- sample(1:NROW(corr$data), 3)
+  si <- sample(seq_len(NROW(corr$data)), 3)
 
   inf <- corr$data[si, c("varx", "vary", "infer.value")]
 
-  for (k in 1:NROW(inf)) {
+  for (k in seq_len(NROW(inf))) {
     expect_equal(m[as.character(inf$vary[k]), as.character(inf$varx[k])], inf$infer.value[k])
   }
 })
