+++ R/corrp.R
@@ -144,12 +144,14 @@ corrp <- function(df,
   on.exit(if (parallel) parallel::stopCluster(cluster))
 
 
-  stopifnot(sapply(df, class) %in% c(
-    "integer",
-    "numeric",
-    "factor",
-    "character"
-  ))
+  
+ 
+  stopifnot(
+    all(
+      vapply(df, class, character(1)) %in% 
+        c("integer", "numeric", "factor", "character")
+    )
+  )
 
 
   cnames <- colnames(df)
