+++ man/corr_matrix.Rd
@@ -29,10 +29,9 @@ create a correlation matrix.
 \examples{
 \dontrun{
 
-air_cor = corrp(airquality)
-air_m = corr_matrix(air_cor,isig = F)
+air_cor <- corrp(airquality)
+air_m <- corr_matrix(air_cor, isig = F)
 corrplot::corrplot(air_m)
-
 }
 
 }
