+++ man/ptest.Rd
@@ -38,8 +38,8 @@ This approximates the null hypothesis — that there is no dependency/difference
 \examples{
 \dontrun{
 
-x = iris[[1]]; y = iris[[2]]
-ptest(x,y,FUN = function(x,y) cor(x,y) ,alternative = 't')
-
+x <- iris[[1]]
+y <- iris[[2]]
+ptest(x, y, FUN = function(x, y) cor(x, y), alternative = "t")
 }
 }
