+++ R/corr_matrix.R
@@ -18,7 +18,7 @@
 #' \dontrun{
 #'
 #' air_cor <- corrp(airquality)
-#' air_m <- corr_matrix(air_cor, isig = F)
+#' air_m <- corr_matrix(air_cor, isig = FALSE)
 #' corrplot::corrplot(air_m)
 #' }
 #'
