+++ R/corr_rm.R
@@ -23,7 +23,7 @@
 #' \dontrun{
 #'
 #' air_cor <- corrp(airquality)
-#' corr_rm(df = airquality, c = air_cor, cutoff = 0.75, col = "infer.value", isig = F)
+#' corr_rm(df = airquality, c = air_cor, cutoff = 0.75, col = "infer.value", isig = FALSE)
 #' }
 #' @export
 corr_rm <- function(df, c, ...) {
