+++ R/ptest.R
@@ -21,7 +21,7 @@
 #'
 ptest <- function(x, y,
                   FUN,
-                  rk = F,
+                  rk = FALSE,
                   alternative = c("two.sided", "less", "greater"),
                   num.s = 1000, ...) {
   FUN <- match.fun(FUN)
