+++ man/corr_matrix.Rd
@@ -30,7 +30,7 @@ create a correlation matrix.
 \dontrun{
 
 air_cor <- corrp(airquality)
-air_m <- corr_matrix(air_cor, isig = F)
+air_m <- corr_matrix(air_cor, isig = FALSE)
 corrplot::corrplot(air_m)
 }
 
