+++ man/corrp.Rd
@@ -11,7 +11,7 @@ corrp(
   p.value = 0.05,
   verbose = TRUE,
   num.s = 1000,
-  rk = F,
+  rk = FALSE,
   comp = c("greater", "less"),
   alternative = c("two.sided", "less", "greater"),
   cor.nn = c("pearson", "mic", "dcor", "pps"),
@@ -130,7 +130,7 @@ categorical variable.
 \dontrun{
 
 air_cor <- corrp(airquality)
-air_m <- corr_matrix(air_cor, isig = F)
+air_m <- corr_matrix(air_cor, isig = FALSE)
 corrplot::corrplot(air_m)
 }
 
