+++ R/acca.R
@@ -5,19 +5,21 @@
 #' @title Average correlation clustering algorithm
 #'
 #' @description A C++ implementation of the ACCA method
-#' that works directly with the correlation 
+#' that works directly with the correlation
 #' matrix derived from the \code{\link{corr_matrix}} function.
 #' In this sense, this implementation differs from the original,
 #' it works with mixed data and several correlation methods.
 #'
-#' @param m  \[\code{matrix(1)}]\cr correlation matrix from 
+#' @param m  \[\code{matrix(1)}]\cr correlation matrix from
 #' \code{\link{corr_matrix}} or a distance matrix.
 #' @param k \[\code{integer(1)}]\cr number of clusters considered.
-#' @param maxrep \[\code{integer(1)}]\cr maximum number of interactions without change in the clusters.
+#' @param maxrep \[\code{integer(1)}]\cr maximum number of
+#' interactions without change in the clusters.
 #' @param maxiter \[\code{integer(1)}]\cr maximum number of interactions.
 #' @param ... Additional arguments (TODO).
 #'
-#' @return \[\code{acca_list(k)}]\cr A list with the final result of the clustering method.
+#' @return \[\code{acca_list(k)}]\cr A list with the
+#' final result of the clustering method.
 #'  That is, the name of the variables belonging to each cluster k.
 #'
 #' @author Igor D.S. Siciliani
@@ -65,7 +67,10 @@ acca.cmatrix <- function(m, k, maxrep = 2L, maxiter = 100L, ...) {
 #' @export
 #' @rdname acca
 acca.matrix <- function(m, k, maxrep = 2L, maxiter = 100L, ...) {
-  warning("m is not an object of the 'cmatrix' class some results may go wrong.")
+  warning(
+    "m is not an object of the 'cmatrix'
+    class some results may go wrong."
+  )
 
 
   k <- as.integer(k)

[…]
