+++ NEWS.md
@@ -9,8 +9,8 @@
   - `Astar`: Derivate the modified distance covariance statistics.
   - `dist`: Calculate distance matrix.
   - `bcdcor`: Function to calculate bias corrected distance correlation.
-  - `dcort`: Function to calculate the t statistic for corrected high dimention distance correlation.
-  - `dcort_test`: Function to calculate the t statistic for corrected high dimention distance correlation. 
+  - `dcort`: Function to calculate the t statistic for corrected high dimension distance correlation.
+  - `dcort_test`: Function to calculate the t statistic for corrected high dimension distance correlation. 
   Returns a list:
     -  method: description of test.
     -  statistic: observed value of the test statistic.
