+++ README.md
@@ -36,7 +36,7 @@ In this new package the correlation is automatically computed according to the f
 - [Predictive Power Score](https://github.com/paulvanderlaken/ppsr).
 
 
-Also, all statistical tests are controlled by the sigficance of
+Also, all statistical tests are controlled by the significance of
 p.value param. If the statistical tests do not obtain a significance greater/less
 than p.value, by default the output of variable `isig` will be `FALSE`.
 There is no statistical significance test for the `pps` algorithm, `isig = TRUE` in this case.
