+++ R/corr_fun.R
@@ -103,11 +103,6 @@
 #' Paul van der Laken, ppsr,2021.
 #' URL \url{https://github.com/paulvanderlaken/ppsr}.
 #'
-#' @examples
-#' \dontrun{
-#'
-#' corr_fun(iris, nx = "Sepal.Length", ny = "Sepal.Width", cor.nn = "dcor")
-#' }
 #'
 #' @export
 corr_fun <- function(df,
