### 2024-10-20 c8d8d9e
feat - improve test coverage #12
