### 2024-10-20 d0be6ad
Add package website and longer form documentation #10 make package long form documentation
