+++ _pkgdown.yml
@@ -1,0 +1,4 @@
+url: https://meantrix.github.io/corrp/
+template:
+  bootstrap: 5
+
