+++ pkgdown.yml
@@ -1,0 +1,9 @@
+pandoc: 2.19.2
+pkgdown: 2.0.6
+pkgdown_sha: ~
+articles: {}
+last_built: 2024-10-20T20:39Z
+urls:
+  reference: https://meantrix.github.io/corrp/reference
+  article: https://meantrix.github.io/corrp/articles
+
