### 2024-10-20 2b2dd79
docs - update package version and changelog #8
