### 2024-10-20 0a32035
Add package website and longer form documentation #10 make vignette
