### 2024-10-28 ae0144d
Document performance and benchmarks #13 Make bench mark scripts
