### 2024-10-30 d3e9e4a
Document performance and benchmarks #13 make full bench mark results
