### 2024-10-30 3b76b26
Document performance and benchmarks #13 update paper with performance improvements
