### 2024-11-01 56725b5
Document performance and benchmarks #13 make statments of benchmark and choices
