+++ paper/paper.md
@@ -68,40 +68,51 @@ The `corrp` package provides seven main functions for correlation calculations, 
 - **sil_acca**: A C++ implementation of the Silhouette method for interpreting and validating the consistency of clusters within ACCA clusters of data.
 - **best_acca**: Determining the optimal number of clusters in ACCA clustering using the average silhouette approach.
 
-First, we calculate the correlations for the *iris* dataset using the Maximal Information Coefficient for numeric pairs, the Predictive Power Score algorithm for numeric/categorical pairs, and the Uncertainty Coefficient for categorical pairs.
+We calculate correlations for the *eusilc* dataset using the Maximal Information Coefficient for numeric pairs, Predictive Power Score for numeric/categorical pairs, and Uncertainty Coefficient for categorical pairs. This synthetic dataset represents Austrian EU-SILC data on income, demographics, and household characteristics.
 
 ```r
+library("laeken")
 library("corrp")
+data(eusilc)
+
+
+eusilc = eusilc[, c("eqSS", "eqIncome", "db040", "rb090")]
+colnames(eusilc) = c("House_Size", "Income", "State", "Sex")
+
 results = corrp(
-  iris, 
-  cor.nn = 'mic', cor.nc = 'pps', cor.cc = 'uncoef', 
-  n.cores = 2, verbose = FALSE
+  eusilc, 
+  cor.nn = 'dcor', cor.nc = 'lm', cor.cc = 'pps',
+  verbose = FALSE
 )
 
 head(results$data)
 ```
 
-|      | infer                           | infer.value | stat       | stat.value |
-|------|---------------------------------|-------------|------------|------------|
-| 1    | Maximal Information Coefficient | 0.9994870   | P-value    | 0.0000000  |
-| 2    | Maximal Information Coefficient | 0.2770503   | P-value    | 0.0000000  |
-| 3    | Maximal Information Coefficient | 0.7682996   | P-value    | 0.0000000  |
-| 4    | Maximal Information Coefficient | 0.6683281   | P-value    | 0.0000000  |
-| 5    | Predictive Power Score          | 0.5591864   | F1_weighted| 0.7028029  |
-| 6    | Maximal Information Coefficient | 0.2770503   | P-value    | 0.0000000  |
 
-
-|      | isig  | msg   | varx         | vary         |
-|------|-------|-------|--------------|--------------|
-| 1    | TRUE  |       | Sepal.Length | Sepal.Length |
-| 2    | TRUE  |       | Sepal.Length | Sepal.Width  |
-| 3    | TRUE  |       | Sepal.Length | Petal.Length |
-| 4    | TRUE  |       | Sepal.Length | Petal.Width  |
-| 5    | TRUE  |       | Sepal.Length | Species      |
-| 6    | TRUE  |       | Sepal.Width  | Sepal.Length |
+|      | infer                  | infer.value    | stat          | stat.value      |
+|------|------------------------|----------------|---------------|-----------------|
+| 1    | Pearson Correlation    | 1.00000000     | P-value       | 0.0000000000    |
+| 2    | Pearson Correlation    | -0.02750826    | P-value       | 0.0008083282    |
+| 3    | Predictive Power Score | 0.00000000     | F1_weighted   | 0.0451575746    |
+| 4    | Predictive Power Score | 0.01811129     | F1_weighted   | 0.4983966359    |
+| 5    | Pearson Correlation    | -0.02750826    | P-value       | 0.0008083282    |
+| 6    | Pearson Correlation    | 1.00000000     | P-value       | 0.0000000000    |
 
 
 
+|      | isig | msg | varx         | vary         |
[…]
