+++ R/ptest.R
@@ -17,7 +17,7 @@ ptest <- function(x, y,
                   FUN,
                   rk = FALSE,
                   alternative = c("two.sided", "less", "greater"),
-                  num.s = 1000, ...) {
+                  num.s = 250, ...) {
   FUN <- match.fun(FUN)
   # check mandatory args
   fargs <- formals(FUN)
