+++ R/corr_fun.R
@@ -110,7 +110,7 @@ corr_fun <- function(df,
                      ny,
                      p.value = 0.05,
                      verbose = TRUE,
-                     num.s = 1000,
+                     num.s = 250,
                      rk = FALSE,
                      comp = c("greater", "less"),
                      alternative = c("two.sided", "less", "greater"),
