+++ man/ptest.Rd
@@ -10,7 +10,7 @@ ptest(
   FUN,
   rk = FALSE,
   alternative = c("two.sided", "less", "greater"),
-  num.s = 1000,
+  num.s = 250,
   ...
 )
 }
