### 2024-11-06 meantrix/corrp#10 @malcolmbarrett
It looks like your vignette is not being picked up by pkgdown. You might need to add it to the yml file

https://pkgdown.r-lib.org/reference/build_articles.html#index-and-navbar
