### 2024-12-08 meantrix/corrp#25 @moylettsinead
corr_rm function has undocumented required parameter 'c'
openjournals/joss-reviews#7319

While testing the package's core functions, corr_rm fails with an unclear error about a missing parameter 'c', which isn't documented.

Required parameter 'c' is not documented. No indication of what values 'c' accepts. Error message doesn't explain what 'c' represents. Not mentioned in README.

Update function documentation to explain:
What the 'c' parameter represents
Valid values for 'c'
Add examples showing different uses of 'c' parameter
