### 2024-12-08 meantrix/corrp#25 @moylettsinead
It seems to be that the matrix from corr_matrix has identified variables as chr and then corr_rm won't work on objects of class "character"
