### 2024-12-08 meantrix/corrp#26 @moylettsinead
best_acca function missing required parameter and lacks documentation
openjournals/joss-reviews#7319

Similar issue as before: Required parameter 'mink' is not documented and README doesn't mention this parameter. Error message doesn't guide users to correct usage.
