### 2024-12-08 meantrix/corrp#27 @moylettsinead
General comments for review
openjournals/joss-reviews#7319

**Functionality**
I've opened a few issues in terms of functionality. Additionally for functionality documentation, it could be more detailed. It's very limited.

**Performance**
There's no specific performance claims noted.
There's limited test coverage in the automated tests.
Missing CONTRIBUTING.md and clear support guidelines (no clear issue templates, limited guidance for support). I found this also when I was testing functionality.

**Software paper**
The statement of need could be strengthened, with a clear indication of what makes the package unique or different to existing packages. Explaining unique features and improvements would be helpful.
The provided examples could be more real-world based to help with interpretability.
