### 2025-01-07 4c8d40c
Document performance and benchmarks #13 update desc
