+++ DESCRIPTION
@@ -35,7 +35,7 @@ License: GPL (>= 3)
 Encoding: UTF-8
 LazyData: true
 LinkingTo: Rcpp, RcppArmadillo
-RoxygenNote: 7.2.3
+RoxygenNote: 7.3.2
 URL: https://github.com/meantrix/corrp,
     https://meantrix.github.io/corrp/
 BugReports: https://github.com/meantrix/corrp/issues
