### 2025-01-08 c9d363c
Document performance and benchmarks #13 fix astar calc
