+++ dev/Rcpp/dist.R
@@ -1,28 +1,40 @@
+## Check consistency with energy
 
-# Test the distCpp function
-set.seed(42)
-gc()
-test_data = readRDS("dev/Rcpp/x.rds") 
-test_data = as.matrix(data.frame(a = c(3, 2, 4, 4)))
+R_Astar = function(d) {
+  if (inherits(d, "dist"))
+    d <- as.matrix(d)
+  n <- nrow(d)
+  if (n != ncol(d)) stop("Argument d should be distance")
+  m <- rowMeans(d)
+  M <- mean(d)
+  a <- sweep(d, 1, m)
+  b <- sweep(a, 2, m)
+  A <- b + M
+
+  A <- A - d / n
+  diag(A) <- m - M
+  (n / (n - 1)) * A
+}
+
 
 Rcpp::sourceCpp("src/dcort.cpp")
+
+
+# test_data = readRDS("dev/Rcpp/x.rds")
+test_data = as.matrix(data.frame(a = c(1, 2, 4, 4)))
+
+
 # Compute distances using distCpp
-dist_r = as.matrix(dist(test_data)) 
-dist
-all(dist_r == dist_cpp)
+r_dist = as.matrix(stats::dist(test_data))
+cpp_dist = dist(test_data)
+all(r_dist == cpp_dist) # TRUE
 
-dist
+r_astar = R_Astar(r_dist)
+cpp_astar = Astar(cpp_dist)
 
-all(Astar(dist_cpp) == Astarcpp(dist_cpp))
+all(r_astar == cpp_astar)
 
-
-
-
-res_cpp = dcorTcpp(as.matrix(data.frame(a = c(3, 2, 4, 4))), as.matrix(data.frame(b = c(5, 3, 3, 7))))
+res_cpp = dcorT_test(as.matrix(data.frame(a = c(3, 2, 4, 4))), as.matrix(data.frame(b = c(5, 3, 3, 7))))
 
 res_energy = energy::dcorT.test(as.matrix(data.frame(a = c(3, 2, 4, 4))), as.matrix(data.frame(b = c(5, 3, 3, 7))))
-class(res_energy) = "list"
 
-dcorT_test(as.matrix(test_data), as.matrix(test_data))
-
-energy::dcorT(as.matrix(test_data), as.matrix(test_data))
