### 2025-01-08 2c4101d
Document performance and benchmarks #13 update message
