### 2025-01-08 87de5e9
Document performance and benchmarks #13 update corrp documentation
