+++ R/corrp.R
@@ -39,7 +39,8 @@
 #'   p.value param. If the statistical tests do not obtain a significance greater/less
 #'   than p.value the value of variable `isig` will be `FALSE`.\cr
 #' - There is no statistical significance test for the pps algorithm. By default `isig` is TRUE.\cr
-#' - If any errors occur during operations the association measure(`infer.value`) will be `NA`.
+#' - If any errors occur during operations the association measure(`infer.value`) will be `NA`.\cr
+#' - The result `data` and `index` will have \eqn{N^2} rows, where N is the number of variables of the input data.
 #'
 #' @param df \[\code{data.frame(1)}]\cr input data frame.
 #' @param parallel \[\code{logical(1)}]\cr If its TRUE run the operations in parallel backend.
