### 2025-01-08 4240c0c
Document performance and benchmarks #13 update doc with example
