+++ man/corr_fun.Rd
@@ -137,6 +137,12 @@ using \code{\link[ppsr]{score}} function.
 The value lies between 0 and 1.\cr
 }
 
+\examples{
+\dontrun{
+# since both `nx` and `ny` columns are numerical the method type is defined by `cor.nn`
+corr_fun(iris, nx = "Sepal.Length", ny = "Sepal.Width", cor.nn = "dcor")
+}
+}
 \references{
 KS Srikanth,sidekicks,cor2, 2020.
 URL \url{https://github.com/talegari/sidekicks/}.
