### 2025-01-08 6513712
General comments for review #27 make documentation update fix vignette on desc
