+++ R/corr_matrix.R
@@ -13,8 +13,11 @@
 #'
 #' @keywords correlation matrix , corrp
 #'
+#' @examples
 #'
-#'
+#' iris_cor <- corrp(iris)
+#' iris_m <- corr_matrix(iris_cor, isig = FALSE)
+#' corrplot::corrplot(iris_m)
 #' @export
 corr_matrix <- function(c, ...) {
   UseMethod("corr_matrix", c)
