+++ R/ptest.R
@@ -9,7 +9,13 @@
 #' @param rk \[\code{logical(1)}]\cr if its TRUE transform x, y numeric vectors with samples ranks.
 #' @param alternative \[\code{character(1)}]\cr a character string specifying the alternative hypothesis,
 #' must be one of "two.sided" (default), "greater" or "less". You can specify just the initial letter.
-#' @param ... Additional arguments (TODO).
+#' @param ... Additional arguments.
+#' 
+#' @examples
+#'
+#' x <- iris[[1]]
+#' y <- iris[[2]]
+#' ptest(x, y, FUN = function(x, y) cor(x, y), alternative = "t")
 #' 
 #' @export
 #'
