+++ R/best_acca.R
@@ -12,7 +12,7 @@
 #' without change in the clusters in the ACCA method.
 #' @param maxiter \[\code{integer(1)}]\cr maximum number
 #' of interactions in the ACCA method.
-#' @param ... Additional arguments (TODO).
+#' @param ... Additional arguments.
 #'
 #' @return \[\code{list(3)}]\cr A list with:
 #' silhouette average with per k `$silhouette.ave`;
@@ -35,7 +35,11 @@
 #' "International Conference on Artificial Intelligence
 #' and Soft Computing. Springer, Cham, 2015.
 #'
+#' @examples
 #'
+#' x <- corrp::corrp(iris)
+#' m <- corrp::corr_matrix(x)
+#' best_acca(m, 2, 6)
 #'
 #' @export
 #'
