+++ R/corr_fun.R
@@ -104,10 +104,10 @@
 #' URL \url{https://github.com/paulvanderlaken/ppsr}.
 #'
 #' @examples
-#' \dontrun{
+#' 
 #' # since both `nx` and `ny` columns are numerical the method type is defined by `cor.nn`
 #' corr_fun(iris, nx = "Sepal.Length", ny = "Sepal.Width", cor.nn = "dcor")
-#' }
+#' 
 #' @export
 corr_fun <- function(df,
                      nx,
