+++ man/corr_rm.Rd
@@ -52,6 +52,12 @@ The default values is 0.75.}
 Remove highly correlated variables from a data.frame using the
 corrp functions outputs and the caret package function \code{\link[caret]{findCorrelation}}.
 }
+\examples{
+
+iris_cor <- corrp(iris)
+corr_rm(df = iris, c = iris_cor, cutoff = 0.75, col = "infer.value", isig = FALSE)
+
+}
 \author{
 Igor D.S. Siciliani
 }
