+++ R/corrp.R
@@ -94,7 +94,7 @@
 #' @examples
 #'  iris_c <- corrp(iris)
 #'  iris_m <- corr_matrix(iris_c, isig = FALSE)
-#'  corrplot::corrplot(iris_m)
+#'  corrplot(iris_m)
 #' 
 #'
 #' @export
