+++ man/corr_matrix.Rd
@@ -30,7 +30,7 @@ create a correlation matrix.
 
 iris_cor <- corrp(iris)
 iris_m <- corr_matrix(iris_cor, isig = FALSE)
-corrplot::corrplot(iris_m)
+corrplot(iris_m)
 }
 \author{
 Igor D.S. Siciliani
