+++ README.md
@@ -65,7 +65,10 @@ remotes::install_github("meantrix/corrp@main")
 `corrp` Next, we calculate the correlations for the data set iris using: Maximal Information Coefficient for numeric pair, the Power Predictive Score algorithm for numeric/categorical pair and Uncertainty coefficient for categorical pair.
 
 ```r
-results = corrp::corrp(iris, cor.nn = 'mic',cor.nc = 'pps',cor.cc = 'uncoef', n.cores = 2 , verbose = FALSE)
+# coorp with using iris using parallel processing
+results = corrp::corrp(iris, cor.nn = 'mic', cor.nc = 'pps',cor.cc = 'uncoef', n.cores = 2 , verbose = FALSE)
+# an sequential example with different correlation pair types
+results_2 = corrp::corrp(mtcars, cor.nn = 'pps', cor.nc = 'lm', cor.cc = 'cramersV', verbose = FALSE)
 
 head(results$data)
 #                            infer infer.value        stat stat.value isig msg         varx         vary
@@ -75,6 +78,16 @@ head(results$data)
 # Maximal Information Coefficient   0.6683281     P-value  0.0000000 TRUE     Sepal.Length  Petal.Width
 #          Predictive Power Score   0.5591864 F1_weighted  0.7028029 TRUE     Sepal.Length      Species
 # Maximal Information Coefficient   0.2770503     P-value  0.0000000 TRUE      Sepal.Width Sepal.Length
+
+head(results_2$data)
+
+#                  infer infer.value stat stat.value isig msg varx vary
+# Predictive Power Score   1.0000000 <NA>         NA TRUE      mpg  mpg
+# Predictive Power Score   0.3861810  MAE  0.8899206 TRUE      mpg  cyl
+# Predictive Power Score   0.3141056  MAE 74.7816795 TRUE      mpg disp
+# Predictive Power Score   0.2311418  MAE 42.3961506 TRUE      mpg   hp
+# Predictive Power Score   0.1646116  MAE  0.3992651 TRUE      mpg drat
+# Predictive Power Score   0.2075760  MAE  0.5768637 TRUE      mpg   wt
 
 ```
 
