+++ NAMESPACE
@@ -23,4 +23,5 @@ export(ptest)
 export(sil_acca)
 importFrom(Rcpp,evalCpp)
 importFrom(RcppArmadillo,armadillo_version)
+importFrom(corrplot,corrplot)
 useDynLib(corrp, .registration=TRUE)
