### 2025-01-09 1da1215
General comments for review #27 update error message
