+++ R/ptest.R
@@ -40,7 +40,7 @@ ptest <- function(x, y,
   checkmate::assert_number(num.s)
   if (is.data.frame(x)) x <- x[[1]]
   if (is.data.frame(y)) y <- y[[1]]
-  if (!rk) stopifnot(is.numeric(x), is.numeric(y))
+  # if (!rk) stopifnot(is.numeric(x), is.numeric(y))
   stopifnot(length(x) == length(y))
 
 
@@ -63,8 +63,8 @@ ptest <- function(x, y,
     "g" = {
       p.value <- mean(est >= obs)
     },
-    "t" = {
-      p.value <- mean(abs(est) >= abs(obs))
+    "t" = {      
+      p.value <- min(mean(est >= obs), mean(est <= obs)) * 2
     }
   )
 
