### 2025-01-10 32c9668
General comments for review #27 make assert_required_argument
