+++ R/corrp.R
@@ -122,7 +122,7 @@ corrp <- function(df,
                   num.s = 250,
                   rk = FALSE,
                   comp = c("greater", "less"),
-                  alternative = c("two.sided", "less", "greater"),
+                  alternative = c("greater", "less", "two.sided"),
                   cor.nn = c("pearson", "mic", "dcor", "pps"),
                   cor.nc = c("lm", "pps"),
                   cor.cc = c("cramersV", "uncoef", "pps"),
