+++ man/ptest.Rd
@@ -9,7 +9,7 @@ ptest(
   y,
   FUN,
   rk = FALSE,
-  alternative = c("two.sided", "less", "greater"),
+  alternative = c("greater", "less", "two.sided"),
   num.s = 250,
   ...
 )
@@ -24,7 +24,7 @@ ptest(
 \item{rk}{[\code{logical(1)}]\cr if its TRUE transform x, y numeric vectors with samples ranks.}
 
 \item{alternative}{[\code{character(1)}]\cr a character string specifying the alternative hypothesis,
-must be one of "two.sided" (default), "greater" or "less". You can specify just the initial letter.}
+must be one of "greater" (default), "less" or "two.sided". You can specify just the initial letter.}
 
 \item{num.s}{[\code{numeric(1)}]\cr number of samples with replacement created with y numeric vector.}
 
