+++ NAMESPACE
@@ -13,6 +13,7 @@ S3method(corr_rm,matrix)
 S3method(sil_acca,acca_list)
 S3method(sil_acca,list)
 export(acca)
+export(assert_required_argument)
 export(best_acca)
 export(corr_fun)
 export(corr_matrix)
