### 2025-01-10 ca2369a
General comments for review #27 fix assert when lazy eval
