+++ R/corr_matrix.R
@@ -3,7 +3,7 @@
 #' @description Through the results obtained from corrp function
 #' create a correlation matrix.
 #'
-#' @param c \[\code{corrp.list(1)}]\cr output from the \code{\link{corrp}} function.
+#' @param c \[\code{clist(1)}]\cr output from the \code{\link{corrp}} function.
 #' @param col \[\code{character(1)}]\cr choose the column to be used in the correlation matrix.
 #' @param isig \[\code{logical(1)}]\cr values that are not statistically significant will
 #' be represented by NA or FALSE in the correlation matrix.
@@ -20,6 +20,7 @@
 #' corrplot(iris_m)
 #' @export
 corr_matrix <- function(c, ...) {
+  assert_required_argument(c, "The 'c' argument must be a clist object, which is the output from corrp.")
   UseMethod("corr_matrix", c)
 }
 
