+++ R/corrp.R
@@ -54,7 +54,7 @@
 #' All statistical tests are controlled by the confidence internal of
 #'   p.value param. If the statistical tests do not obtain a significance greater/less
 #'   than p.value the value of variable `isig` will be `FALSE`.\cr
-#' There is no statistical significance test for the pps algorithm. By default `isig` is TRUE.\cr
+#' By default there is no statistical significance test for the pps algorithm. By default `isig` is NA, you can enable in the pps.args.\cr
 #' If any errors occur during operations the association measure (`infer.value`) will be `NA`.\cr
 #' The result `data` and `index` will have \eqn{N^2} rows, where N is the number of variables of the input data.
 #'
@@ -130,7 +130,7 @@ corrp <- function(df,
                   pearson.args = list(),
                   dcor.args = list(),
                   mic.args = list(),
-                  pps.args = list(),
+                  pps.args = list(ptest = FALSE),
                   cramersV.args = list(),
                   uncoef.args = list(),
                   ...) {
