+++ R/corrp.R
@@ -8,29 +8,25 @@
 #'
 #' @name corrp
 #'
-#' @section Details (Pair Types):
+#' @section Pair Types:
 #'
-#' - \code{integer/numeric pair} Pearson Correlation using \code{\link[stats]{cor}} function. The
-#'   value lies between -1 and 1.\cr
-#' - \code{integer/numeric pair} Distance Correlation using \code{\link[energy]{dcorT.test}} function. The
-#'   value lies between 0 and 1.\cr
-#' - \code{integer/numeric pair} Maximal Information Coefficient using \code{\link[minerva]{mine}} function. The
-#'   value lies between 0 and 1.\cr
-#' - \code{integer/numeric pair} Predictive Power Score using \code{\link[ppsr]{score}} function. The
-#'   value lies between 0 and 1.\cr\cr
-#' - \code{integer/numeric - factor/categorical pair} correlation coefficient or
-#'   squared root of R^2 coefficient of linear regression of integer/numeric
-#'   variable over factor/categorical variable using \code{\link[stats]{lm}} function. The value
-#'   lies between 0 and 1.\cr
-#' - \code{integer/numeric - factor/categorical pair} Predictive Power Score using \code{\link[ppsr]{score}} function. The
-#'   value lies between 0 and 1.\cr\cr
-#' - \code{factor/categorical pair} Cramer's V value is
-#'   computed based on chisq test and using \code{\link[lsr]{cramersV}} function. The value lies
-#'   between 0 and 1.\cr
-#' - \code{factor/categorical pair} Uncertainty coefficient using \code{\link[DescTools]{UncertCoef}} function. The
-#'   value lies between 0 and 1.\cr
-#' - \code{factor/categorical pair} Predictive Power Score using \code{\link[ppsr]{score}} function. The
-#'   value lies between 0 and 1.\cr
+#' **Numeric pairs (integer/numeric):**
+#' 
+#' - **Pearson Correlation Coefficient:** A widely used measure of the strength and direction of linear relationships. Implemented using \code{\link[stats]{cor}}. For more details, see \url{https://doi.org/10.1098/rspl.1895.0041}. The value lies between -1 and 1.\cr
+#' - **Distance Correlation:** Based on the idea of expanding covariance to distances, it measures both linear and nonlinear associations between variables. Implemented using \code{\link[energy]{dcorT.test}}. For more details, see \url{https://doi.org/10.1214/009053607000000505}. The value lies between 0 and 1.\cr
+#' - **Maximal Information Coefficient (MIC):** An information-based nonparametric method that can detect both linear and non-linear relationships between variables. Implemented using \code{\link[minerva]{mine}}. For more details, see \url{https://doi.org/10.1126/science.1205438}. The value lies between 0 and 1.\cr
+#' - **Predictive Power Score (PPS):** A metric used to assess predictive relations between variables. Implemented using \code{\link[ppsr]{score}}. For more details, see \url{https://zenodo.org/record/4091345}. The value lies between 0 and 1.\cr\cr
+#'
+#' **Numeric and categorical pairs (integer/numeric - factor/categorical):**
+#' 
+#' - **Square Root of R² Coefficient:** From linear regression of the numeric variable over the categorical variable. Implemented using \code{\link[stats]{lm}}. For more details, see \url{https://doi.org/10.4324/9780203774441}. The value lies between 0 and 1.\cr
[…]
