+++ R/sil_acca.R
@@ -4,7 +4,7 @@
 #' of interpretation and validation of consistency within acca clusters of data.
 #'
 #' @param acca \[\code{acca_list(1)}]\cr Acca clustering results from \code{\link{acca}}
-#' @param m  \[\code{matrix(1)}]\cr correlation matrix from \code{\link{corr_matrix}}.
+#' @param m  \[\code{cmatrix(1)|matrix(1)}]\cr correlation matrix from \code{\link{corr_matrix}}.
 #' By default the distance matrix(dist) used in this method is given by `dist = 1 - m`.
 #' @param ... Additional arguments.
 #'
@@ -36,7 +36,9 @@
 #'
 #' @export
 #'
-sil_acca <- function(acca, ...) {
+sil_acca <- function(acca, m, ...) {
+  assert_required_argument(acca, "The 'acca' argument must be a acca_list object, which is the output from acca function, or it must be a list.")
+  assert_required_argument(m, "The 'm' argument must be a cmatrix object, which is the output from corr_matrix function or it must be a matrix.")
   UseMethod("sil_acca", acca)
 }
 
