### 2025-01-10 meantrix/corrp#27 @PHS-Meantrix
The PR #15 has addressed all the issues you opened. However, I have a few questions to clarify some points:

1. Could you please explain what you mean by "truncated message on different methods"? Is this related to an RStudio visualization feature?
2. The performance claims are included in the paper.  Is there a more suitable place to include them?

Thank you for your feedback!
