### 2025-01-23 #7319 @jromanowska
@PhilBoileau, this is mainly R package, and you are not supposed to be expert in the C++ implementation to be able to review that. Could you please take a look at [reviewer checklist](https://joss.readthedocs.io/en/latest/review_checklist.html) and think about it once more?
