### 2025-01-23 #7319 @PhilBoileau
@jromanowska, a third of this package’s code is written in C++. I can’t vouch for the functionality of this software if I can’t review a non-negligible portion of its code.
