### 2025-02-05 meantrix/corrp#27 @devSJR
Hello @PHS-Meantrix, I am trying to help with the review of openjournals/joss-reviews#7319 now. Regarding the performance, there is juste the statement that corrp uses the foreach package to significantly improve performance on large datasets. Though reasonable to believe, there is no performance analysis for this claim. At least I was unable to spot it. My suggestion is to either add it to the paper (which would lengthen it) or to put the results in the README or preferably to add a vignette.
