### 2025-02-05 #7319 @devSJR
I have just checked some of the latest commits, and it appears that @malcolmbarrett already spotted many issues that I also saw. It appears that the issues like the statement about the performance we substantiated by measurements. I will ask the bot for an updated PDF.
