### 2025-02-05 meantrix/corrp#28 @devSJR
Grammar and spelling
Dear corrp team,

please go through your package and check for grammar and spelling mistakes to complete openjournals/joss-reviews#7319.

I found for example

'# auxiliar functions' in utils

and '"#'@description Execute one-sample permutation test on two numeric vector.' in ptest.R

Maybe integrate the spelling package (https://cran.r-project.org/package=spelling).
