### 2025-02-05 meantrix/corrp#29 @devSJR
Add the most traditional correlation methods in R
Please consider for openjournals/joss-reviews#7319 adding the names of the most traditional correlation methods applicable in R to the corresponding section in paper.md
