### 2025-02-05 meantrix/corrp#30 @devSJR
Not run
I usually discourage including examples as 'Not run'

air_cor = corrp(airquality)
air_m = corr_matrix(air_cor,isig = F)
corrplot::corrplot(air_m)

In this example, users might get the error:

Error in loadNamespace(x) : there is no package called ‘corrplot’

openjournals/joss-reviews#7319
