### 2025-02-05 meantrix/corrp#31 @devSJR
Code style
Most of the code looks styled well. However, there seem to be some misses, like in the example of corr_rm

air_cor = corrp(airquality)
corr_rm(df=airquality,c=air_cor,cutoff=0.75,col = 'infer.value',isig = F)

I recommend running code styling over all code for openjournals/joss-reviews#7319
