### 2025-02-05 meantrix/corrp#32 @devSJR
CRAN
Dear authors, during the review for openjournals/joss-reviews#7319, I was wondering if there are plans to put corrp on CRAN?
