### 2025-02-07 meantrix/corrp#34 @devSJR
Remarks regarding JOSS paper
Dear corrp authors,

due to technical difficulties, the second reviewer @anspiess asked me to add some of his remarks for the review for openjournals/joss-reviews#7319

Here are his remarks:

1. SUMMARY => Please explain in more detail for what scientific questions the ACCA algorithm is used for, not just “genetic studies”, and what its advantages are in comparison to other clustering methods. Also mention that it is a supervised method since k needs to be defined.
2. STATEMENT OF NEED => This section is much too short. The authors should be more elaborate on what “mixed-type data” means and when it arises (for instance, the iris dataset). In the provided package, the corrp function automatically detects the underlying data type (Q1: is this failproof?), but in the real world, it is not widespread knowledge that the normal cor function works with continuous-continuous and continuous-ordinal (has order) data, and that for continuous-nominal (no order) point-biserial correlation and for nominal-nominal Cramer’s V needs to be used. Please be more explicit on the basics here as to provide the reader with some background when each method is employed.
3. Line 23: missing full stop after “datasets”.
4. IMPLEMENTATION: Since the corresponding method is selected automatically based on the data-type pair, it would be better to explicitly state “automated selection” in line 30 (see Q1).
5. One thing puzzles me: When running the airquality (6 variables) example in the corrp function help, I get an output of 36 pairwise correlation tests = 6^2. This implies that not only pair a-b is tested but also pair b-a, which only occurs when the data type is not automatically identified, otherwise we would only have pair a-b (a is continuous, b is categorical) and thus (6^2)/2 – 6 tests (minus 6 because the diagonal of the correlation matrix is always 1. Please comment on what actually happens under the hood, automated selection or plain pairwise combinatorial testing.
6. Along the same lines: It is not clear how “significance” is deduced in the output. For instance, in “Solar.R vs. Ozone . Alternative hypothesis: true Pearson Correlation  is not equal to 0. P-value:  0.000179310857164885“, was the p-value multiple testing-corrected (Bonferroni, FDR etc)? Essentially, in correlation matrix you are conducting n^2 statistical tests, which need to be corrected for, compare the corr.test function in the psych package.
7. USAGE: bullet point cor_rm => Please explain in 1-2 sentences how the variables are removed by p-values. Successively, iteratively, all at once, what is the purpose?

9. Comments on the package

Good: I installed the package from GitHub and installation/C++ compilation passed.
    1. Please correct the packages README.md file according to your corrections in the manuscript.
    2. I noticed that the package uses testthat to ensure that the results conform to those of other established packages or base cor. This is good!
    3. I am not a fan of using x = 3 in R assignment, when the common syntax is x <- 3, although it works. At least in the function examples, use <- for assignment.
    4. In general, the help documentation is quite rudimentary. Please be more verbose in the “Descriptions” and add a few more “Examples”.
    5. Concerning “Examples”, it would be advisable to remove the /dontruns, because the code is not thoroughly checked (only if the -run-dontrun flag is activated).
    6. The parameters under “Arguments” and “Values” are not depicted correct, for instance in ptest:
x	\[numeric(1)]
a numeric vector.
All these errors can be alleviated by changing
\item{x}{\[\code{numeric(1)}]\cr a numeric vector.} => \item{x}{a numeric vector.}
    7. Don’t write \item{...}{Additional arguments (TODO).} but \item{...}{Additional future arguments.}
