### 2025-02-16 a9a44bd
fix- set version 0.6.0 in documents #34
