+++ R/corr_rm.R
@@ -1,18 +1,16 @@
-#' @title corr_rm filter data.frame by correlation values.
+#' @title Filter Data Frame by Correlation Values
 #'
-#' @description Remove highly correlated variables from a data.frame using the
-#' corrp functions outputs and the caret package function \code{\link[caret]{findCorrelation}}.
+#' @description Remove highly correlated variables from a data frame using the
+#' correlation functions' outputs and the \code{\link[caret]{findCorrelation}} function from the caret package.
 #'
-#'
-#' @param df \[\code{data.frame(1)}]\cr input data frame.
-#' @param c \[\code{clist(1)} | \code{cmatrix(1)}]\cr correlation list output from the function \code{\link[corrp]{corrp}}
-#' with class \code{clist} or correlation matrix output
-#' from \code{\link[corrp]{corr_matrix}} with class \code{cmatrix}.
-#' @param cutoff \[\code{numeric(1)}]\cr A numeric value for the pair-wise absolute correlation cutoff.
-#' The default values is 0.75.
-#' @param col \[\code{character(1)}]\cr choose the column to be used in the correlation matrix
-#' @param isig \[\code{logical(1)}]\cr values that are not statistically significant will
-#' be represented by NA or FALSE in the correlation matrix.
+#' @param df \[\code{data.frame(1)}]\cr The input data frame.
+#' @param c \[\code{clist(1)} | \code{cmatrix(1)}]\cr A correlation list output from the \code{\link[corrp]{corrp}} function (with class \code{clist}), 
+#' or a correlation matrix output from the \code{\link[corrp]{corr_matrix}} function (with class \code{cmatrix}).
+#' @param cutoff \[\code{numeric(1)}]\cr A numeric value for the pairwise absolute correlation cutoff. 
+#' The default value is 0.75.
+#' @param col \[\code{character(1)}]\cr The column to be used in the correlation matrix.
+#' @param isig \[\code{logical(1)}]\cr Whether values that are not statistically significant should 
+#' be represented by \code{NA} or \code{FALSE} in the correlation matrix.
 #' @param ... Additional arguments.
 #'
 #' @examples
@@ -24,11 +22,11 @@
 #'
 #' @author Igor D.S. Siciliani, Paulo H. dos Santos
 #'
-#' @keywords highly correlated , cmatrix , clist
+#' @keywords highly correlated, cmatrix, clist
 #'
 #' @export
 corr_rm <- function(df, c, ...) {
-  assert_required_argument(df, "The 'df' argument must be a data.frame which columns will be filtered.")
+  assert_required_argument(df, "The 'df' argument must be a data frame from which columns will be filtered.")
   assert_required_argument(c, "The 'c' argument must be a clist object, which is the output from corrp, or a cmatrix object, which is the output from corr_matrix.")
   UseMethod("corr_rm", c)
 }
@@ -49,7 +47,7 @@ corr_rm.clist <- function(df, c, col = c("infer.value", "stat.value"),
 #' @rdname corr_rm
 corr_rm.list <- function(df, c, col = c("infer.value", "stat.value"),
                          isig = TRUE, cutoff = 0.75, ...) {
-  warning("it is not an object of the 'clist' class some results may go wrong.")
+  warning("This is not an object of the 'clist' class; some results may be incorrect.")
 
   col <- match.arg(col)
 
@@ -71,7 +69,7 @@ corr_rm.cmatrix <- function(df, c, cutoff = 0.75, ...) {
 #' @export
 #' @rdname corr_rm
 corr_rm.matrix <- function(df, c, cutoff = 0.75, ...) {
[…]
