### 2025-02-18 19f4f94
Remarks regarding JOSS paper #34 update corr_matrix
