### 2025-02-18 993074f
Remarks regarding JOSS paper #34 adress ...
