### 2025-02-18 1a8762a
Remarks regarding JOSS paper #34 make a lot of example for corrp
