+++ R/corr_matrix.R
@@ -17,7 +17,7 @@
 #'
 #' iris_cor <- corrp(iris)
 #' iris_m <- corr_matrix(iris_cor, isig = FALSE)
-#' corrplot::corrplot(iris_m)
+#' corrplot(iris_m)
 #' @export
 corr_matrix <- function(c, ...) {
   assert_required_argument(c, "The 'c' argument must be a clist object, which is the output from corrp.")
@@ -67,3 +67,7 @@ corr_matrix.clist <- function(c, col = c("infer.value", "stat.value", "isig"), i
 
   return(structure(m, class = c("cmatrix", "matrix")))
 }
+
+#' @inherit corrplot::corrplot
+#' @export
+corrplot <- corrplot::corrplot
