+++ NAMESPACE
@@ -19,11 +19,11 @@ export(corr_fun)
 export(corr_matrix)
 export(corr_rm)
 export(corrp)
+export(corrplot)
 export(dcorT_test)
 export(ptest)
 export(set_arguments)
 export(sil_acca)
 importFrom(Rcpp,evalCpp)
 importFrom(RcppArmadillo,armadillo_version)
-importFrom(corrplot,corrplot)
 useDynLib(corrp, .registration=TRUE)
