### 2025-02-18 ba58fa8
Remarks regarding JOSS paper #34 fix checks
