+++ man/corrplot.Rd
@@ -1,559 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/corr_matrix.R
-\name{corrplot}
-\alias{corrplot}
-\title{A visualization of a correlation matrix.}
-\usage{
-corrplot(
-  corr,
-  method = c("circle", "square", "ellipse", "number", "shade", "color", "pie"),
-  type = c("full", "lower", "upper"),
-  col = NULL,
-  col.lim = NULL,
-  is.corr = TRUE,
-  bg = "white",
-  title = "",
-  add = FALSE,
-  diag = TRUE,
-  outline = FALSE,
-  mar = c(0, 0, 0, 0),
-  addgrid.col = NULL,
-  addCoef.col = NULL,
-  addCoefasPercent = FALSE,
-  order = c("original", "AOE", "FPC", "hclust", "alphabet"),
-  hclust.method = c("complete", "ward", "ward.D", "ward.D2", "single", "average",
-    "mcquitty", "median", "centroid"),
-  addrect = NULL,
-  rect.col = "black",
-  rect.lwd = 2,
-  tl.pos = NULL,
-  tl.cex = 1,
-  tl.col = "red",
-  tl.offset = 0.4,
-  tl.srt = 90,
-  cl.pos = NULL,
-  cl.length = NULL,
-  cl.cex = 0.8,
-  cl.ratio = 0.15,
-  cl.align.text = "c",
-  cl.offset = 0.5,
-  number.cex = 1,
-  number.font = 2,
-  number.digits = NULL,
-  addshade = c("negative", "positive", "all"),
-  shade.lwd = 1,
-  shade.col = "white",
-  transKeepSign = TRUE,
-  p.mat = NULL,
-  sig.level = 0.05,
-  insig = c("pch", "p-value", "blank", "n", "label_sig"),
-  pch = 4,
-  pch.col = "black",
-  pch.cex = 3,
-  plotCI = c("n", "square", "circle", "rect"),
-  lowCI.mat = NULL,
-  uppCI.mat = NULL,
-  na.label = "?",
-  na.label.col = "black",
-  win.asp = 1,
-  ...
-)
-}
-\arguments{
-\item{corr}{The correlation matrix to visualize, must be square if
-\code{order} is not \code{'original'}. For general matrix, please using
-\code{is.corr = FALSE} to convert.}
-
-\item{method}{Character, the visualization method of correlation matrix to be
-  used. Currently, it supports seven methods, named \code{'circle'}
-  (default), \code{'square'}, \code{'ellipse'}, \code{'number'},
-  \code{'pie'}, \code{'shade'} and \code{'color'}. See examples for details.
-
-  The areas of circles or squares show the absolute value of corresponding
-  correlation coefficients. Method \code{'pie'} and \code{'shade'} came from
-  Michael Friendly's job (with some adjustment about the shade added on), and
-  \code{'ellipse'} came from D.J. Murdoch and E.D. Chow's job, see in section
-  References.}
-
-\item{type}{Character, \code{'full'} (default), \code{'upper'} or
-\code{'lower'}, display full matrix, lower triangular or upper triangular
-matrix.}
-
-\item{col}{Vector, the colors of glyphs. They are distributed uniformly in
-\code{col.lim} interval.
-If \code{is.corr} is \code{TRUE}, the default value will be \code{COL2('RdBu', 200)}.
-If \code{is.corr} is \code{FALSE} and \code{corr} is a non-negative or non-positive matrix,
-the default value will be \code{COL1('YlOrBr', 200)};
-otherwise (elements are partly positive and partly negative),
-the default value will be \code{COL2('RdBu', 200)}.}
-
-\item{col.lim}{The limits \code{(x1, x2)} interval for assigning color by
[…]
