### 2025-02-18 471f063
Remarks regarding JOSS paper #34 update reame and  corrp documentation sync with corr_fun
