### 2025-02-18 ae9765b
Remarks regarding JOSS paper #34 finish acca documentation
