+++ R/acca.R
@@ -6,11 +6,9 @@
 
 #' @title Average correlation clustering algorithm
 #'
-#' @description A C++ implementation of the ACCA method
-#' that works directly with the correlation
-#' matrix derived from the \code{\link{corr_matrix}} function.
-#' In this sense, this implementation differs from the original,
-#' it works with mixed data and several correlation methods.
+#' @description A C++ implementation of the Average Correlation Clustering Algorithm (ACCA) \url{https://www.sciencedirect.com/science/article/pii/S1532046410000158}, originally developed for genetic studies using Pearson correlation as a similarity measure. Unlike traditional clustering methods that rely on distance metrics such as Euclidean or Mahalanobis distance, ACCA groups data based on correlation patterns.\cr
+#' This implementation works directly with the correlation matrix derived from the \code{\link{corr_matrix}} function and supports mixed data types along with various correlation methods.\cr
+#' ACCA is an unsupervised clustering method, meaning it identifies patterns without predefined labels. Similar to k-means, it requires defining the K parameter, which controls the number of clusters.
 #'
 #' @param m  \[\code{matrix(1)}]\cr correlation matrix from
 #' \code{\link{corr_matrix}} or a distance matrix.
@@ -22,7 +20,7 @@
 #'
 #' @return \[\code{acca_list(k)}]\cr A list with the
 #' final result of the clustering method.
-#'  That is, the name of the variables belonging to each cluster k.
+#'  That is, every element of the list group names of the variables belonging to each cluster k.
 #'
 #' @author Igor D.S. Siciliani, Paulo H. dos Santos
 #'
@@ -36,9 +34,25 @@
 #'
 #' @examples
 #'
+#' # Clustering a correlation matrix with 3 clusters
 #' x <- corrp::corrp(iris)
 #' m <- corrp::corr_matrix(x)
-#' corrp::acca(m, 2)
+#' result <- corrp::acca(m, k = 3)
+#' print(result)
+#'
+#' # Clustering with 5 clusters and increasing the maximum number of interactions
+#' x <- corrp::corrp(iris)
+#' m <- corrp::corr_matrix(x)
+#' result <- corrp::acca(m, k = 5, maxiter = 200)
+#' print(result)
+#'
+#'
+#' # Adjusting the maximum number of iterations without change in clusters
+#' x <- corrp::corrp(iris)
+#' m <- corrp::corr_matrix(x)
+#' result <- corrp::acca(m, k = 2, maxrep = 50)
+#' print(result)
+#'
 #'
 #' @export
 #'
