### 2025-02-18 281be7b
Remarks regarding JOSS paper #34 finish style
