### 2025-02-18 a469e30
Remarks regarding JOSS paper #34 update news
