### 2025-02-18 80f93e6
Remarks regarding JOSS paper #34 update news 2
