### 2025-02-19 e7f1e02
Add the most traditional correlation methods in R #29 make the cor method map
