### 2025-02-19 c9aeb47
Add the most traditional correlation methods in R #29 fix minor doc prob
