### 2025-02-19 meantrix/corrp#34 @PHS-Meantrix
@anspiess, @devSJR, PR #35 has addressed all the issues you raised. Regarding subitem 6 of item 8 in @anspiess's request, we would like to maintain our current documentation style. This approach quickly shows the user the class and length of each argument in an organized manner. The style is based on the 'checkmate' package.
