### 2025-03-05 #7319 @devSJR
> [@anspiess](https://github.com/anspiess), you have mentioned by email to me that you were waiting for the authors to address some comments. I'm not sure what comments you mean - there are no comments from you here in this thread.

It refers to this issue https://github.com/meantrix/corrp/issues/34#issue-2837860751
