### 2025-03-05 #7319 @jromanowska
> > [@anspiess](https://github.com/anspiess), you have mentioned by email to me that you were waiting for the authors to address some comments. I'm not sure what comments you mean - there are no comments from you here in this thread.
>
> It refers to this issue [meantrix/corrp#34 (comment)](https://github.com/meantrix/corrp/issues/34#issue-2837860751)

Thanks for clarification! @anspiess - could you comment here or in the other issue?
