### 2025-03-25 82655e6
Update paper.md
suggestions to improve paper

Signed-off-by: Julia <23313172+jromanowska@users.noreply.github.com>
