### 2025-03-25 f4bff72
Update paper.md
more corrections and suggestions

Signed-off-by: Julia <23313172+jromanowska@users.noreply.github.com>
