### 2025-03-26 846c525
Update README.md
fix language mistakes and correct some formulations to make the text easier to read

Signed-off-by: Julia <23313172+jromanowska@users.noreply.github.com>
