+++ README.md
@@ -7,18 +7,18 @@
 
 <!-- badges: end -->
 
-Compute multiple types of correlation analysis (Pearson correlation, R^2  coefficient of linear regression, Cramer's V measure of association, Distance Correlation, The Maximal Information Coefficient, Uncertainty coefficient and Predictive Power Score) in large dataframes with mixed columns classes(integer, numeric, factor and character) in parallel R backend.
+Compute multiple types of correlation analysis (Pearson correlation, R^2  coefficient of linear regression, Cramer's V measure of association, Distance Correlation, The Maximal Information Coefficient, Uncertainty coefficient and Predictive Power Score) in large dataframes with mixed columns classes (integer, numeric, factor and character) in parallel R backend.
+
 This package also has a C++ implementation of the Average correlation clustering algorithm [ACCA](https://www.sciencedirect.com/science/article/pii/S1532046410000158) 
-that works directly with the correlation matrix. In this sense, this implementation differs from the original,
-it works with mixed data and several correlation types methods.
+that is adapted to work directly with the correlation matrix. In this sense, this implementation works with mixed data and several correlation methods.
 
 ## Details
 
-The [corrp package](https://github.com/meantrix/corrP) under development by Meantrix team and original based on Srikanth KS (talegari) [cor2 function](https://github.com/talegari/sidekicks/) can provide to R users a way to work with correlation analysis among large data.frames, tibbles or data.tables through a R parallel backend and C++ functions.
+The [corrp package](https://github.com/meantrix/corrP) is under development by Meantrix team and based on Srikanth KS (`@talegari`) [cor2 function](https://github.com/talegari/sidekicks/). The package enables R users to perform a correlation analysis on large data frames (`data.frame`, `tibble`, or `data.tables`) through a R parallel backend and C++ functions.
 
-The data.frame is allowed to have columns of these four classes: integer, numeric, factor and character. The character column is considered as categorical variable.
+The input data frame is allowed to have columns of these four classes: integer, numeric, factor, and character. The character column is considered as a categorical variable.
 
-In this new package the correlation is automatically computed according to the follow options: 
+`{corr}` can compute correlation on the following pairs: 
 
 #### integer/numeric pair:
 - [Pearson correlation test](https://en.wikipedia.org/wiki/Pearson_correlation_coefficient) ;
@@ -36,10 +36,11 @@ In this new package the correlation is automatically computed according to the f
 - [Predictive Power Score](https://github.com/paulvanderlaken/ppsr).
 
 
-Also, all statistical tests are controlled by the sigficance of
-p.value param. If the statistical tests do not obtain a significance greater/less
-than p.value, by default the output of variable `isig` will be `FALSE`.
+Also, all statistical tests are performed and the results can then be assessed for their sigficance, when setting the
+`p.value` parameter. If the statistical tests do not obtain a significance greater/less
+than `p.value`, by default the output of variable `isig` will be `FALSE`.
 There is no statistical significance test for the `pps` algorithm, `isig = TRUE` in this case.
+
 If any errors occur during operations by default the correlation will be `NA`.
 
 
@@ -62,10 +63,10 @@ remotes::install_github("meantrix/corrp@main")
[…]
