### 2025-03-26 e7a1911
fix - paper table
https://github.com/meantrix/corrp/pull/37#discussion_r2012892348
