### 2025-03-26 c233841
fix - mention some R methods and provide refs
https://github.com/meantrix/corrp/pull/37#discussion_r2012887915
