+++ paper/paper.bib
@@ -79,3 +79,21 @@ doi = {10.1016/j.jbi.2010.02.001},
 url = {https://www.sciencedirect.com/science/article/pii/S1532046410000158},
 author = {Anindya Bhattacharya and Rajat K. De}
 }
+
+
+@Manual{stats:2024,
+  title        = {The R Stats Package: Statistical Functions},
+  author       = {{R Core Team}},
+  year         = {2024},
+  note         = {R package version 4.6.0},
+  url          = {https://stat.ethz.ch/R-manual/R-devel/library/stats/html/00Index.html}
+}
+
+@Manual{Hmisc:2025,
+  title        = {Hmisc: Harrell Miscellaneous},
+  author       = {Frank E. Harrell Jr},
+  year         = {2025},
+  note         = {R package version 5.2-3},
+  url          = {https://cran.r-project.org/package=Hmisc}
+}
+
