### 2025-03-26 0358e32
fix - add sidekicks talegari ref
