+++ paper/paper.bib
@@ -97,3 +97,11 @@ author = {Anindya Bhattacharya and Rajat K. De}
   url          = {https://cran.r-project.org/package=Hmisc}
 }
 
+@Misc{sidekicks:2025,
+  author       = {Srikanth Komala Sheshachala},
+  title        = {sidekicks: A Misc Set of Functions for Data Analysis},
+  year         = {2025},
+  howpublished = {\url{https://github.com/talegari/sidekicks}},
+  note         = {Accessed: 2025-03-26}
+}
+
