+++ paper/paper.bib
@@ -105,3 +105,14 @@ author = {Anindya Bhattacharya and Rajat K. De}
   note         = {Accessed: 2025-03-26}
 }
 
+
+@Manual{laeken:2024,
+  title        = {laeken: Estimation of indicators on social exclusion and poverty, as well as Pareto tail modeling for empirical income distributions},
+  author       = {Andreas Alfons and Matthias Templ and Peter Filzmoser},
+  year         = {2024},
+  note         = {R package version 0.5.3},
+  url          = {https://CRAN.R-project.org/package=laeken}
+}
+
+
+
