+++ pkgdown.yml
@@ -1,9 +1,9 @@
-pandoc: 2.19.2
-pkgdown: 2.0.6
+pandoc: 3.1.11
+pkgdown: 2.1.1
 pkgdown_sha: ~
-articles: {}
-last_built: 2024-10-20T20:39Z
+articles:
+  usage-corrp: usage-corrp.html
+last_built: 2025-03-26T20:15Z
 urls:
   reference: https://meantrix.github.io/corrp/reference
   article: https://meantrix.github.io/corrp/articles
-
