### 2025-03-31 #7319 @jromanowska
@igor-siciliani , @PHS-Meantrix
I've been checking the examples in README and here are some points you should address:

[ ] use [lintr](https://lintr.r-lib.org/) and/or [styler](https://styler.r-lib.org/) on the code (and on the entire source code of the package)
[ ] be consistent: either you use `corrp::` in front of each function from the package or you start the examples by loading the package (`library(corrp)`) - when getting through the README code, I couldn't run `corr_matrix` because suddenly it was without the `corrp::`
[ ] there are many warning messages (`In (function (df, x, y, algorithm = "tree", metrics = list(regression = "MAE",  ... : There are on average only 6.4 observations in each test-set for the mpg-cyl relationship.`) when executing `corrp` on `mtcars` due to small number of observations in this dataset - please consider a dataset that's larger, e.g., [palmer penguins](https://allisonhorst.github.io/palmerpenguins/)
[ ] perhaps add a visualization of the correlation matrix?
