### 2025-03-31 #7319 @jromanowska
I see also that you have a lot of unresolved issues from the previous reviewers. Could you go through those and check which ones can be resolved? I will wait with reviewing further until you have completed some of the tasks they mentioned. Let me know when you're done :)
