+++ man/dcorT_test.Rd
@@ -1,10 +1,10 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/ptest.R
-\name{dcorT_test}
-\alias{dcorT_test}
+\name{dcor_t_test}
+\alias{dcor_t_test}
 \title{Create Correlation Matrix from corrp inferences}
 \usage{
-dcorT_test(x, y)
+dcor_t_test(x, y)
 }
 \arguments{
 \item{x}{[\code{data.frame(1) | matrix(1)}]\cr Data of the first sample.}
