+++ NAMESPACE
@@ -19,7 +19,7 @@ export(corr_fun)
 export(corr_matrix)
 export(corr_rm)
 export(corrp)
-export(dcorT_test)
+export(dcor_t_test)
 export(ptest)
 export(set_arguments)
 export(sil_acca)
