### 2025-03-31 04ffbcd
Error in tests #41 fix when only installed
