+++ README.md
@@ -73,7 +73,8 @@ remotes::install_github("meantrix/corrp@main")
 # coorp with using iris using parallel processing
 results = corrp::corrp(iris, cor.nn = 'mic', cor.nc = 'pps',cor.cc = 'uncoef', n.cores = 2 , verbose = FALSE)
 # an sequential example with different correlation pair types
-results_2 = corrp::corrp(mtcars, cor.nn = 'pps', cor.nc = 'lm', cor.cc = 'cramersV', parallel = FALSE, verbose = FALSE)
+data(penguins, package = 'palmerpenguins')
+results_2 = corrp::corrp(penguins, cor.nn = 'pps', cor.nc = 'lm', cor.cc = 'cramersV', parallel = FALSE, verbose = FALSE)
 
 head(results$data)
 #                            infer infer.value        stat stat.value isig msg         varx         vary
@@ -86,13 +87,15 @@ head(results$data)
 
 head(results_2$data)
 
-#                  infer infer.value stat stat.value isig msg varx vary
-# Predictive Power Score   1.0000000 <NA>         NA NA      mpg  mpg
-# Predictive Power Score   0.3861810  MAE  0.8899206 NA      mpg  cyl
-# Predictive Power Score   0.3141056  MAE 74.7816795 NA      mpg disp
-# Predictive Power Score   0.2311418  MAE 42.3961506 NA      mpg   hp
-# Predictive Power Score   0.1646116  MAE  0.3992651 NA      mpg drat
-# Predictive Power Score   0.2075760  MAE  0.5768637 NA      mpg   wt
+#        infer infer.value    stat    stat.value isig msg    varx                 vary
+#   Cramer's V   1.0000000 P-value  4.997501e-04 TRUE     species              species
+#   Cramer's V   0.6598431 P-value  4.997501e-04 TRUE     species               island
+# Linear Model   0.8413139 P-value  2.694614e-91 TRUE     species       bill_length_mm
+# Linear Model   0.8244751 P-value  1.507658e-84 TRUE     species        bill_depth_mm
+# Linear Model   0.8821728 P-value 1.351710e-111 TRUE     species    flipper_length_mm
+# Linear Model   0.8183349 P-value  2.892368e-82 TRUE     species          body_mass_g
+  
+
 
 ```
 
