### 2025-04-01 meantrix/corrp#42 @PHS-Meantrix
Ptest documentation
> Throughout the documentation, you use `clist` as output and input to some functions - I'm not sure I understand what it is. Is it `list` you ment? Moreover, you're using various ways to describe output of the functions. Please use only one, like in the [`dcorT_test`](https://meantrix.github.io/corrp/reference/dcorT_test.html). And apropos `dcorT_test`:
>
> - [x] it should say "data", not "A data" - data is plural
> - [x] please use `snake_case` throughout, so not `dcorT_test` but `dcor_t_test`
> - [x] add `@return` to [ptest](https://meantrix.github.io/corrp/reference/ptest.html) documentation

 _Originally posted by @jromanowska in [#7319](https://github.com/openjournals/joss-reviews/issues/7319#issuecomment-2765289247)_
