### 2025-04-01 meantrix/corrp#43 @PHS-Meantrix
README and Vignette
> I've been checking the examples in README and here are some points you should address:
>
> - [x] use [lintr](https://lintr.r-lib.org/) and/or [styler](https://styler.r-lib.org/) on the code (and on the entire source code of the package)
> - [x] be consistent: either you use `corrp::` in front of each function from the package or you start the examples by loading the package (`library(corrp)`) - when getting through the README code, I couldn't run `corr_matrix` because suddenly it was without the `corrp::`
> - [x] there are many warning messages (`In (function (df, x, y, algorithm = "tree", metrics = list(regression = "MAE",  ... : There are on average only 6.4 observations in each test-set for the mpg-cyl relationship.`) when executing `corrp` on `mtcars` due to small number of observations in this dataset - please consider a dataset that's larger, e.g., [palmer penguins](https://allisonhorst.github.io/palmerpenguins/)
> - [x] perhaps add a visualization of the correlation matrix?
>
> Further, in the [vignette](https://meantrix.github.io/corrp/articles/usage-corrp.html), please:
>
> - [x] explain the names of the names of the columns in "Exploring the Results"
> - [x] add explanation on how to change the default significance threshold in "Filtering Significant Correlations"
> - [x] add explanation of the results and perhaps comparison with the previous calculation in "Correlation Types"
> - [x] choose another dataset and perhaps use [{bench}](https://bench.r-lib.org/) to show that your package is efficient on large datasets, also in parallel, in "Parallel Processing"
>
> _(do not tick off the fields, I will do that)_

 _Originally posted by @jromanowska in [#7319](https://github.com/openjournals/joss-reviews/issues/7319#issuecomment-2765240898)_
