### 2025-04-01 d504495
README and Vignette #43 style the whole directory remove notes on package
