### 2025-04-01 8ab8c1b
README and Vignette #43 make a phrase for the comparison between the arguments
