### 2025-04-01 ea63d16
README and Vignette #43 update usage-corrp
