### 2025-04-01 438ad6f
README and Vignette #43 update package to have the package without notes
