+++ NAMESPACE
@@ -26,4 +26,5 @@ export(sil_acca)
 importFrom(Rcpp,evalCpp)
 importFrom(RcppArmadillo,armadillo_version)
 importFrom(corrplot,corrplot)
+importFrom(palmerpenguins,path_to_file)
 useDynLib(corrp, .registration=TRUE)
