### 2025-04-01 2791607
README and Vignette #43 make the parallel comparison
