### 2025-04-02 0fcfd25
README and Vignette #43 fix parallel and make bench
