### 2025-04-02 c2cc7ec
README and Vignette #43 fix parallel and make bench
