### 2025-04-02 62a514c
README and Vignette #43 fix rcmdcheck time and notes
