+++ DESCRIPTION
@@ -31,6 +31,7 @@ Imports:
     palmerpenguins (>= 0.1.1)
 Suggests: 
     energy,
+    laeken,
     knitr,
     rmarkdown,
     testthat
@@ -44,3 +45,4 @@ URL: https://github.com/meantrix/corrp,
     https://meantrix.github.io/corrp/
 BugReports: https://github.com/meantrix/corrp/issues
 VignetteBuilder: knitr
+
