### 2025-04-02 ee8e8f3
README and Vignette #43 update package doc
