+++ R/ptest.R
@@ -2,12 +2,12 @@
 #' @description Execute a one-sample permutation test on two numeric vectors.
 #' One vector is kept constant while the other is "shuffled" by resampling.
 #' This approximates the null hypothesis — that there is no dependency or difference between the variables.
-#' @param x \[\code{numeric(1)}]\cr A numeric vector.
-#' @param y \[\code{numeric(1)}]\cr A numeric vector.
-#' @param FUN \[\code{function(1)}]\cr The function to be applied.
-#' @param num.s \[\code{numeric(1)}]\cr The number of samples with replacement created from the y numeric vector.
-#' @param rk \[\code{logical(1)}]\cr If TRUE, transform x and y numeric vectors with sample ranks.
-#' @param alternative \[\code{character(1)}]\cr A character string specifying the alternative hypothesis.
+#' @param x \[\code{numeric(1)}\]\cr A numeric vector.
+#' @param y \[\code{numeric(1)}\]\cr A numeric vector.
+#' @param FUN \[\code{function(1)}\]\cr The function to be applied.
+#' @param num.s \[\code{numeric(1)}\]\cr The number of samples with replacement created from the y numeric vector.
+#' @param rk \[\code{logical(1)}\]\cr If TRUE, transform x and y numeric vectors with sample ranks.
+#' @param alternative \[\code{character(1)}\]\cr A character string specifying the alternative hypothesis.
 #' Must be one of "greater" (default), "less", or "two.sided". You can specify just the initial letter.
 #' @param ... Additional arguments.
 #'
@@ -17,7 +17,7 @@
 #' y <- iris[[2]]
 #' ptest(x, y, FUN = function(x, y) cor(x, y), alternative = "t")
 #'
-#' @return \[\code{numeric(1)}]\cr The p.value statistic.
+#' @return \[\code{numeric(1)}\]\cr The p.value statistic.
 #'
 #' @export
 ptest <- function(x, y,
@@ -77,10 +77,10 @@ ptest <- function(x, y,
 #'
 #' @description Distance correlation t-test of multivariate independence for high dimension. C++ version of energy::dcorT.test.
 #'
-#' @param x \[\code{data.frame(1) | matrix(1)}]\cr Data of the first sample.
-#' @param y \[\code{data.frame(1) | matrix(1)}]\cr Data of the second sample.
+#' @param x \[\code{data.frame(1) | matrix(1)}\]\cr Data of the first sample.
+#' @param y \[\code{data.frame(1) | matrix(1)}\]\cr Data of the second sample.
 #'
-#' @return \[\code{list(6)}]\cr returns a list containing:
+#' @return \[\code{list(6)}\]\cr returns a list containing:
 #'   \item{method}{description of test.}
 #'   \item{statistic}{observed value of the test statistic.}
 #'   \item{parameter}{degrees of freedom.}
