+++ R/corr_rm.R
@@ -3,13 +3,13 @@
 #' @description Remove highly correlated variables from a data frame using the
 #' correlation functions' outputs and the \code{\link[caret]{findCorrelation}} function from the caret package.
 #'
-#' @param df \[\code{data.frame(1)}]\cr The input data frame.
-#' @param c \[\code{clist(1)} | \code{cmatrix(1)}]\cr A correlation list output from the \code{\link[corrp]{corrp}} function (with class \code{clist}),
+#' @param df \[\code{data.frame(1)}\]\cr The input data frame.
+#' @param c \[\code{clist(1)} | \code{cmatrix(1)}\]\cr A correlation list output from the \code{\link[corrp]{corrp}} function (with class \code{clist}),
 #' or a correlation matrix output from the \code{\link[corrp]{corr_matrix}} function (with class \code{cmatrix}).
-#' @param cutoff \[\code{numeric(1)}]\cr A numeric value for the pairwise absolute correlation cutoff.
+#' @param cutoff \[\code{numeric(1)}\]\cr A numeric value for the pairwise absolute correlation cutoff.
 #' The default value is 0.75.
-#' @param col \[\code{character(1)}]\cr The column to be used in the correlation matrix.
-#' @param isig \[\code{logical(1)}]\cr Whether values that are not statistically significant should
+#' @param col \[\code{character(1)}\]\cr The column to be used in the correlation matrix.
+#' @param isig \[\code{logical(1)}\]\cr Whether values that are not statistically significant should
 #' be represented by \code{NA} or \code{FALSE} in the correlation matrix.
 #' @param ... Additional arguments.
 #'

[…]
