### 2025-04-02 6cc7cd1
README and Vignette #43 update package doc
