### 2025-04-02 2501fcb
Add automated `R CMD CHECK` to GitHub Actions #21 make gitactions
