### 2025-04-02 35e284c
Add automated `R CMD CHECK` to GitHub Actions #21 update parallel to git action remove ubuntu_old
