### 2025-04-02 d6a9444
Add automated `R CMD CHECK` to GitHub Actions #21  update to fix for win
