### 2025-04-02 ed90213
Add automated `R CMD CHECK` to GitHub Actions #21  update to fix for win 2
