### 2025-04-02 #7319 @PHS-Meantrix
@jromanowska
I have updated the package and documentation according to the comments:

https://github.com/openjournals/joss-reviews/issues/7319#issuecomment-2765240898
https://github.com/openjournals/joss-reviews/issues/7319#issuecomment-2765289247

Let me know if anything is missing. The unresolved issues were just not closed; the only one that was not addressed is the CRAN submission, as we do not plan to make available through CRAN. I believe you can continue the review process.
