### 2025-04-07 #7319 @jromanowska
@PHS-Meantrix, @igor-siciliani - thank you for the improvements. Although, I was surprised that they are only in JOSS branch. I assume you will merge it to main after we're done?
I was looking through the code and I found lots of tests in the `dev` folder - tried running some in the `acca_test.R` and I couldn't: it cannot find some functions (like `crand_acca` or `subset2d`). Could you explain how to run the tests there? Why didn't you include those tests as part of the automatic testing with `testthat`?
