### 2025-04-24 #7319 @jromanowska
Just a small question: I can see that even though you merged many of the [branches](https://github.com/meantrix/corrp/branches), you still have not deleted them - is there a reason for that?
