### 2025-04-24 #7319 @devSJR
I would also be happy about a less verbose structure. It will make it easier to track everything.
