### 2025-04-24 #7319 @jromanowska
We're almost there, @PHS-Meantrix! Some more comments and questions:

- [ ] I'd like you to extend the tests that are run in testthat. If you run `goodpractice::gp(pkg_dir, checks = "covr")`, you'll see that e.g., `acca.cpp` and `utils.R` have very many lines that are not tested.
- [ ] How would one use [`corr_rm`](https://meantrix.github.io/corrp/reference/corr_rm.html) function? It says that it removes correlated columns - but the correlation between vars A and B, doesn't mean that var A cannot be used with yet another var C, so why should one remove the correlated pairs?
- [ ] In `corr_rm` documentation, there is no information about the return value.
- [ ] More from documentation: it's difficult to understand difference between `corr_fun` and `corrp` - please, change the titles. Do I understand correctly that `corr_fun` is for analysis of single pair of variables, while `corrp` uses `corr_fun` to go through all the pairs in the given dataframe?
- [ ] I don't think that the title of [`dcor_t_test`](https://meantrix.github.io/corrp/reference/dcor_t_test.html) documentation is correct. This doesn't return a correlation matrix, but a list... also: what does "(bias corrected) squared dCor(x,y)" mean?
- [ ] Moreover: there are no examples for  `dcor_t_test` - I'm not sure I understand what it does...
- [ ] In [`ptest`](https://meantrix.github.io/corrp/reference/ptest.html) - why are there "additional arguments"? For future development?
- [ ] Why are these two "utility functions" exported and how would the user use them? https://meantrix.github.io/corrp/reference/set_arguments.html and https://meantrix.github.io/corrp/reference/assert_required_argument.html
