### 2025-04-24 #7319 @jromanowska
Apropos not testing all the lines: I've just wanted to run the `corrp` on a specific dataset I'm working with now and I got this very uninformative error message:
```
Error in 'corrp::corrp(df)' :
all(vapply(df, class, character(1)) %in% c("integer", "numeric", ... is not TRUE
```
I've figured out that's because I had some variables that were 'logical', but in this case, you should check and inform the user that this function doesn't handle 'logical' variables. And you can even choose to remove those and continue with the analysis on the remaining variables.
