### 2025-04-24 #7319 @jromanowska
I've also tried clustering and searching for the best clustering with `best_acca`, but I don't understand these two arguments:

```
maxrep

    [integer(1)]
    maximum number of interactions without change in the clusters in the ACCA method.
maxiter

    [integer(1)]
    maximum number of interactions in the ACCA method.
```
What "interactions" are you talking about here? And what "change in the clusters"?
