### 2025-05-01 #7319 @PHS-Meantrix
https://github.com/openjournals/joss-reviews/issues/7319#issuecomment-2827304175
@jromanowska - We've addressed all the points listed in the closed [PR](https://github.com/meantrix/corrp/pull/48). The review can continue. Let us know if anything else is required.
