### 2025-05-05 #7319 @jromanowska
Thank you, @PHS-Meantrix and @igor-siciliani - great work!
I see that you've removed the `@export` in the two utility functions I've mentioned, but they still render on the webpage - are there some pckgdown arguments you need to tweak? Or maybe re-rendering will do...

Apropos `corr_rm` - I'm still not convinced that it's a useful function. _How would you use it?_

If a user wants to remove the correlated variables, the user needs to decide _which_ variables from the correlated pair to remove. Sometimes, it will be the first one, sometimes the second one - it depends on the research question. From what I understand, your function simply removes the second variable from the correlated pairs.
