### 2025-05-05 #7319 @jromanowska
When I run tests locally, I get some warnings and errors:

```
==> devtools::test()

ℹ Testing corrp
✔ | F W  S  OK | Context
✔ |         11 | corr_matrix
✔ |   2      4 | corr_rm
────────────────────────────────────────────────────────────────────────────────────────────────────────
Warning (test-corr_rm.R:18:3): Tests on corr_rm
This is not an object of the 'clist' class; some results may be incorrect.
Backtrace:
    ▆
 1. ├─corrp::corr_rm(df = df, c = corr2) at test-corr_rm.R:18:3
 2. └─corrp:::corr_rm.list(df = df, c = corr2) at corrp/R/corr_rm.R:36:3

Warning (test-corr_rm.R:18:3): Tests on corr_rm
The provided object is not of class 'clist'; some results may be incorrect.
Backtrace:
    ▆
 1. ├─corrp::corr_rm(df = df, c = corr2) at test-corr_rm.R:18:3
 2. └─corrp:::corr_rm.list(df = df, c = corr2) at corrp/R/corr_rm.R:36:3
 3.   ├─corrp::corr_matrix(c = c, col = col, isig = isig, ...) at corrp/R/corr_rm.R:59:3
 4.   └─corrp:::corr_matrix.default(c = c, col = col, isig = isig, ...) at corrp/R/corr_matrix.R:32:3
────────────────────────────────────────────────────────────────────────────────────────────────────────
⠙ |         82 | corrp                                                                                  Error in if (compare$comp) { : missing value where TRUE/FALSE needed
ERROR: some operations produces Nas values.
 D  FUN  A
Error : Missing values present in input variable 'x'. Consider using use = 'pairwise.complete.obs'.
Error : Missing values present in input variable 'x'. Consider using use = 'pairwise.complete.obs'.
Error : Missing values present in input variable 'x'. Consider using use = 'pairwise.complete.obs'.
Error : Missing values present in input variable 'y'. Consider using use = 'pairwise.complete.obs'.
Error in if (compare$comp) { : missing value where TRUE/FALSE needed
Error : Missing values present in input variable 'x'. Consider using use = 'pairwise.complete.obs'.
Error : Missing values present in input variable 'x'. Consider using use = 'pairwise.complete.obs'.
Note: y was forced from character to factor.
Error : Missing values present in input variable 'x'. Consider using use = 'pairwise.complete.obs'.
Error : Missing values present in input variable 'y'. Consider using use = 'pairwise.complete.obs'.
Error in model.frame.default(Terms, newdata, na.action = na.action, xlev = attr(object,  :
  factor x has new level c
Error : Missing values present in input variable 'x'. Consider using use = 'pairwise.complete.obs'.
Error : Missing values present in input variable 'x'. Consider using use = 'pairwise.complete.obs'.
Note: y was forced from character to factor.
Error : Missing values present in input variable 'x'. Consider using use = 'pairwise.complete.obs'.
Error : Missing values present in input variable 'y'. Consider using use = 'pairwise.complete.obs'.
Error : Missing values present in input variable 'x'. Consider using use = 'pairwise.complete.obs'.
Error : Missing values present in input variable 'x'. Consider using use = 'pairwise.complete.obs'.
Note: y was forced from character to factor.
Error : Missing values present in input variable 'x'. Consider using use = 'pairwise.complete.obs'.
✔ |         88 | corrp [28.9s]
✔ |          4 | ptest
✔ |         13 | tests-acca [1.1s]

══ Results ═════════════════════════════════════════════════════════════════════════════════════════════
Duration: 30.8 s

[ FAIL 0 | WARN 2 | SKIP 0 | PASS 120 ]
```

I've run the `test-corrp.R` manually and it seems like these come from lines 227-235.
