### 2025-05-06 #7319 @PHS-Meantrix
@jromanowska -

> Just a small question: I can see that even though you merged many of the [branches](https://github.com/meantrix/corrp/branches), you still have not deleted them - is there a reason for that?

I've deleted every non-active branch — now only main, gh_pages, and JOSS remain.

> I see that you've removed the @export in the two utility functions I've mentioned, but they still render on the webpage - are there some pckgdown arguments you need to tweak? Or maybe re-rendering will do...

The website has been updated, removing the internal functions from rendering. It is now organized into two sections: Correlation Analysis and Clustering Algorithms.

> the user needs to decide which variables from the correlated pair to remove

It is not advisable to use this function if you want to manually select which column to remove. This function is meant to reduce multicollinearity automatically: when a pair exceeds the correlation threshold, the variable with the largest mean absolute correlation between the pair is removed.

Let me know if anything else is needed.
