### 2025-05-06 #7319 @PHS-Meantrix
https://github.com/openjournals/joss-reviews/issues/7319#issuecomment-2850264236

@jromanowska - The devtools::test() problems were also addressed.
