### 2025-05-07 #7319 @jromanowska
Thanks, but I can still see the same errors when I run the tests - you have not changed the lines I've marked in the comment above.

Apropos `corr_rm` - you say it's not advisable to use it for manual selection of columns to remove. I understand that but on the documentation page, it says "This function is suited for mitigating multicollinearity issues in statistical models." I'm not sure how could one use that - usually, when I construct a model, I decide what is needed and what is not, based on the scientific question I want to answer. While in prediction, it's actually sometimes useful to have many correlated variables.
_How would you use this function?_
