### 2025-05-19 #7319 @jromanowska
> > Before were fixed warning issues, now we fixed all the error messages (not actual test errors). If some message appears install locally and restart the session, since the messages are given by process that can be parallel.
>
> I _had_ installed locally and restarted. I understand that the errors are not test errors but they still _were_ errors.
>
> > > Apropos corr_rm - you say it's not advisable to use it for manual selection ... How would you use this function?
> >
> >
> > One example of this filtering step is found in a [genomic prediction study](https://www.ncbi.nlm.nih.gov/pmc/articles/PMC7938892/), where the authors removed highly correlated variables (SNPs). This was important because, with thousands of SNPs, there is often a lot of redundancy. This preprocessing is useful when the goal is prediction, not interpretation, where the hypotheses about each predictor are a concern.
> > For a kmeans high correlation between can introduce bias toward the correlated variables, this is not a big issue for cforest.
> > A proposed [workflow](https://www.matthewalanham.com/Students/2018_MWDSI_R%20caret%20paper.pdf) by Lanham demonstrates this filtering process as a step before model training. In this paper there is R examples with the carret::findCorrelation, which is the core the corr_rm function.
>
> Great example, thank you! Please, add this explanation to the documentation.

@PHS-Meantrix , @igor-siciliani , have you added this explanation to the documentation?
