### 2025-05-21 #7319 @jromanowska
I've checked the doi (10.1016/0304-4076(75)90038-x) - it's not the correct one, so we'll ignore this suggestion form editorialbot.
