### 2025-05-21 #7319 @PHS-Meantrix
Hi @jromanowska, Here is the information you requested. Let me know if need anything else.

DOI: https://doi.org/10.5281/zenodo.15482741
Version: 0.6.0
