### 2025-05-22 #7319 @editorialbot
:wave: @openjournals/dsais-eics, this paper is ready to be accepted and published.
